test(s3s-fs): migrate it_aws to the s3s-test harness
- enable the s3s-test custom harness for the aws test target (harness = false)
- replace the hand-written #[tokio::test] integration test with suite/fixture/case registration: FsServer suite (standard config) with domain fixtures (Bucket/List/Object/Multipart/Copy/Conditional/Sts) and an FsServerRelaxed suite for the relaxed name validation case
- migrate all 54 cases with identical assertions (177 asserts, 1:1)
- drop the serial() global lock, static config, log_and_unwrap!, and manual tracing setup; cases are isolated by UUID buckets and run concurrently

Refs #394
