+++ crates/s3s-fs/build.rs
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: 2023-2026 The s3s Authors
+
+fn main() {
+    s3s_test::build::collect_info();
+}
