--- crates/s3s-fs/tests/it_aws.rs
@@ -1,3811 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// SPDX-FileCopyrightText: 2023-2026 The s3s Authors
-
-use s3s::auth::SimpleAuth;
-use s3s::header::CONTENT_TYPE;
-use s3s::host::SingleDomain;
-use s3s::service::S3ServiceBuilder;
-use s3s::validation::NameValidation;
-use s3s_fs::FileSystem;
-
-use std::env;
-use std::fs;
-
-use aws_config::SdkConfig;
-use aws_credential_types::provider::SharedCredentialsProvider;
-use aws_sdk_s3::Client;
-use aws_sdk_s3::config::Credentials;
-use aws_sdk_s3::config::Region;
-use aws_sdk_s3::primitives::ByteStream;
-
-use aws_sdk_s3::types::BucketLocationConstraint;
-use aws_sdk_s3::types::ChecksumMode;
-use aws_sdk_s3::types::CompletedMultipartUpload;
-use aws_sdk_s3::types::CompletedPart;
-use aws_sdk_s3::types::CreateBucketConfiguration;
-
-use aws_sdk_s3::error::ProvideErrorMetadata;
-
-use anyhow::Result;
-use hyper::Method;
-use tokio::sync::Mutex;
-use tokio::sync::MutexGuard;
-use tracing::{debug, error};
-use uuid::Uuid;
-
-const FS_ROOT: &str = concat!(env!("CARGO_TARGET_TMPDIR"), "/s3s-fs-tests-aws");
-const DOMAIN_NAME: &str = "localhost:8014";
-const REGION: &str = "us-west-2";
-
-// STS AssumeRole route that returns NotImplemented
-struct AssumeRoleRoute;
-
-#[async_trait::async_trait]
-impl s3s::route::S3Route for AssumeRoleRoute {
-    fn is_match(&self, method: &Method, uri: &hyper::Uri, headers: &hyper::HeaderMap, _: &mut hyper::http::Extensions) -> bool {
-        if method == Method::POST
-            && uri.path() == "/"
-            && let Some(val) = headers.get(CONTENT_TYPE)
-            && val.as_bytes() == b"application/x-www-form-urlencoded"
-        {
-            return true;
-        }
-        false
-    }
-
-    async fn call(&self, _req: s3s::S3Request<s3s::Body>) -> s3s::S3Result<s3s::S3Response<s3s::Body>> {
-        debug!("AssumeRole called - returning NotImplemented");
-        Err(s3s::s3_error!(NotImplemented, "STS operations are not supported by s3s-fs"))
-    }
-}
-
-fn setup_tracing() {
-    use tracing_subscriber::EnvFilter;
-
-    // if env::var("RUST_LOG").is_err() {
-    //     // TODO: Audit that the environment access only happens in single-threaded code.
-    //     unsafe { env::set_var("RUST_LOG", "it_aws=debug,s3s_fs=debug,s3s=debug") };
-    // }
-
-    tracing_subscriber::fmt()
-        .pretty()
-        .with_env_filter(EnvFilter::from_default_env())
-        .with_test_writer()
-        .init();
-}
-
-fn config() -> &'static SdkConfig {
-    use std::sync::LazyLock;
-    static CONFIG: LazyLock<SdkConfig> = LazyLock::new(|| {
-        setup_tracing();
-
-        // Fake credentials
-        let cred = Credentials::for_tests();
-
-        // Setup S3 provider
-        fs::create_dir_all(FS_ROOT).unwrap();
-        let fs = FileSystem::new(FS_ROOT).unwrap();
-
-        // Setup S3 service
-        let service = {
-            let mut b = S3ServiceBuilder::new(fs);
-            b.set_auth(SimpleAuth::from_single(cred.access_key_id(), cred.secret_access_key()));
-            b.set_host(SingleDomain::new(DOMAIN_NAME).unwrap());
-            b.set_route(AssumeRoleRoute);
-            b.build()
-        };
-
-        // Convert to aws http client
-        let client = s3s_aws::Client::from(service);
-
-        // Setup aws sdk config
-        SdkConfig::builder()
-            .credentials_provider(SharedCredentialsProvider::new(cred))
-            .http_client(client)
-            .region(Region::new(REGION))
-            .endpoint_url(format!("http://{DOMAIN_NAME}"))
-            .build()
-    });
-    &CONFIG
-}
-
-async fn serial() -> MutexGuard<'static, ()> {
-    use std::sync::LazyLock;
-    static LOCK: LazyLock<Mutex<()>> = LazyLock::new(|| Mutex::new(()));
-    LOCK.lock().await
-}
-
-fn create_client_with_validation(validation: impl NameValidation + 'static) -> Client {
-    // Setup with custom validation
-    let service = {
-        let fs = FileSystem::new(FS_ROOT).unwrap();
-        let mut b = S3ServiceBuilder::new(fs);
-        let cred = Credentials::for_tests();
-        b.set_auth(SimpleAuth::from_single(cred.access_key_id(), cred.secret_access_key()));
-        b.set_host(SingleDomain::new(DOMAIN_NAME).unwrap());
-        b.set_validation(validation);
-        b.build()
-    };
-
-    // Convert to aws http client
-    let client_inner = s3s_aws::Client::from(service);
-
-    // Setup aws sdk config
-    let cred = Credentials::for_tests();
-    let config = SdkConfig::builder()
-        .credentials_provider(SharedCredentialsProvider::new(cred))
-        .http_client(client_inner)
-        .region(Region::new(REGION))
-        .endpoint_url(format!("http://{DOMAIN_NAME}"))
-        .build();
-
-    Client::new(&config)
-}
-
-async fn create_bucket(c: &Client, bucket: &str) -> Result<()> {
-    let location = BucketLocationConstraint::from(REGION);
-    let cfg = CreateBucketConfiguration::builder().location_constraint(location).build();
-
-    c.create_bucket()
-        .create_bucket_configuration(cfg)
-        .bucket(bucket)
-        .send()
-        .await?;
-
-    debug!("created bucket: {bucket:?}");
-    Ok(())
-}
-
-async fn delete_object(c: &Client, bucket: &str, key: &str) -> Result<()> {
-    c.delete_object().bucket(bucket).key(key).send().await?;
-    Ok(())
-}
-
-async fn delete_bucket(c: &Client, bucket: &str) -> Result<()> {
-    c.delete_bucket().bucket(bucket).send().await?;
-    Ok(())
-}
-
-async fn do_multipart_upload(c: &Client, bucket: &str, key: &str, content: &[u8]) -> Result<(String, Vec<CompletedPart>)> {
-    let upload_id = c
-        .create_multipart_upload()
-        .bucket(bucket)
-        .key(key)
-        .send()
-        .await?
-        .upload_id
-        .ok_or_else(|| anyhow::anyhow!("create_multipart_upload response missing upload_id"))?;
-
-    let ans = c
-        .upload_part()
-        .bucket(bucket)
-        .key(key)
-        .upload_id(&upload_id)
-        .body(ByteStream::from(content.to_vec()))
-        .part_number(1)
-        .send()
-        .await?;
-
-    let e_tag = ans.e_tag.ok_or_else(|| anyhow::anyhow!("upload_part returned no ETag"))?;
-
-    let part = CompletedPart::builder().e_tag(e_tag).part_number(1).build();
-
-    Ok((upload_id, vec![part]))
-}
-
-macro_rules! log_and_unwrap {
-    ($result:expr) => {
-        match $result {
-            Ok(ans) => {
-                debug!(?ans);
-                ans
-            }
-            Err(err) => {
-                error!(?err);
-                return Err(err.into());
-            }
-        }
-    };
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_list_buckets() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let response1 = log_and_unwrap!(c.list_buckets().send().await);
-    drop(response1);
-
-    let bucket1 = format!("test-list-buckets-1-{}", Uuid::new_v4());
-    let bucket1_str = bucket1.as_str();
-    let bucket2 = format!("test-list-buckets-2-{}", Uuid::new_v4());
-    let bucket2_str = bucket2.as_str();
-
-    create_bucket(&c, bucket1_str).await?;
-    create_bucket(&c, bucket2_str).await?;
-
-    let response2 = log_and_unwrap!(c.list_buckets().send().await);
-    let bucket_names: Vec<_> = response2.buckets().iter().filter_map(|bucket| bucket.name()).collect();
-    assert!(bucket_names.contains(&bucket1_str));
-    assert!(bucket_names.contains(&bucket2_str));
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_list_objects_v2() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-list-objects-v2-{}", Uuid::new_v4());
-    let bucket_str = bucket.as_str();
-    create_bucket(&c, bucket_str).await?;
-
-    let test_prefix = "/this/is/a/test/";
-    let key1 = "this/is/a/test/path/file1.txt";
-    let key2 = "this/is/a/test/path/file2.txt";
-    {
-        let content = "hello world\nनमस्ते दुनिया\n";
-        let crc32c = base64_simd::STANDARD.encode_to_string(crc32c::crc32c(content.as_bytes()).to_be_bytes());
-        c.put_object()
-            .bucket(bucket_str)
-            .key(key1)
-            .body(ByteStream::from_static(content.as_bytes()))
-            .checksum_crc32_c(crc32c.as_str())
-            .send()
-            .await?;
-        c.put_object()
-            .bucket(bucket_str)
-            .key(key2)
-            .body(ByteStream::from_static(content.as_bytes()))
-            .checksum_crc32_c(crc32c.as_str())
-            .send()
-            .await?;
-    }
-
-    let result = c.list_objects_v2().bucket(bucket_str).prefix(test_prefix).send().await;
-
-    let response = log_and_unwrap!(result);
-
-    let contents: Vec<_> = response.contents().iter().filter_map(|obj| obj.key()).collect();
-    assert_ne!(contents.len(), 0);
-    assert!(contents.contains(&key1));
-    assert!(contents.contains(&key2));
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_list_objects_v2_with_prefixes() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-list-prefixes-{}", Uuid::new_v4());
-    let bucket_str = bucket.as_str();
-    create_bucket(&c, bucket_str).await?;
-
-    // Create files in nested directory structure
-    let content = "hello world\n";
-    let files = [
-        "README.md",                   // Root level file
-        "test/subdirectory/README.md", // Nested file
-        "test/file.txt",               // File in test/ directory
-        "other/dir/file.txt",          // File in other/dir/ directory
-    ];
-
-    for key in &files {
-        c.put_object()
-            .bucket(bucket_str)
-            .key(*key)
-            .body(ByteStream::from_static(content.as_bytes()))
-            .send()
-            .await?;
-    }
-
-    // List without delimiter - should return all files recursively
-    let result = c.list_objects_v2().bucket(bucket_str).send().await;
-
-    let response = log_and_unwrap!(result);
-    let contents: Vec<_> = response.contents().iter().filter_map(|obj| obj.key()).collect();
-
-    debug!("List without delimiter - objects: {:?}", contents);
-    assert_eq!(contents.len(), 4);
-    for key in &files {
-        assert!(contents.contains(key), "Missing key: {key}");
-    }
-
-    // List with delimiter "/" - should return root files and common prefixes
-    let result = c.list_objects_v2().bucket(bucket_str).delimiter("/").send().await;
-
-    let response = log_and_unwrap!(result);
-
-    // Should have one file at root level
-    let contents: Vec<_> = response.contents().iter().filter_map(|obj| obj.key()).collect();
-    debug!("List with delimiter - objects: {:?}", contents);
-    assert_eq!(contents.len(), 1);
-    assert!(contents.contains(&"README.md"));
-
-    // Should have two common prefixes: "test/" and "other/"
-    let prefixes: Vec<_> = response.common_prefixes().iter().filter_map(|cp| cp.prefix()).collect();
-    debug!("List with delimiter - prefixes: {:?}", prefixes);
-    assert_eq!(prefixes.len(), 2);
-    assert!(prefixes.contains(&"test/"));
-    assert!(prefixes.contains(&"other/"));
-
-    // List with prefix "test/" and delimiter "/" - should return files in test/ and subdirectories
-    let result = c
-        .list_objects_v2()
-        .bucket(bucket_str)
-        .prefix("test/")
-        .delimiter("/")
-        .send()
-        .await;
-
-    let response = log_and_unwrap!(result);
-
-    // Should have one file in test/ directory
-    let contents: Vec<_> = response.contents().iter().filter_map(|obj| obj.key()).collect();
-    debug!("List with prefix test/ - objects: {:?}", contents);
-    assert_eq!(contents.len(), 1);
-    assert!(contents.contains(&"test/file.txt"));
-
-    // Should have one common prefix: "test/subdirectory/"
-    let prefixes: Vec<_> = response.common_prefixes().iter().filter_map(|cp| cp.prefix()).collect();
-    debug!("List with prefix test/ - prefixes: {:?}", prefixes);
-    assert_eq!(prefixes.len(), 1);
-    assert!(prefixes.contains(&"test/subdirectory/"));
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_list_objects_v1_with_prefixes() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-list-v1-prefixes-{}", Uuid::new_v4());
-    let bucket_str = bucket.as_str();
-    create_bucket(&c, bucket_str).await?;
-
-    // Create a simple structure
-    let content = "hello world\n";
-    let files = ["README.md", "dir/file.txt"];
-
-    for key in &files {
-        c.put_object()
-            .bucket(bucket_str)
-            .key(*key)
-            .body(ByteStream::from_static(content.as_bytes()))
-            .send()
-            .await?;
-    }
-
-    // Test list_objects (v1) with delimiter
-    let result = c.list_objects().bucket(bucket_str).delimiter("/").send().await;
-
-    let response = log_and_unwrap!(result);
-
-    // Should have one file at root level
-    let contents: Vec<_> = response.contents().iter().filter_map(|obj| obj.key()).collect();
-    debug!("ListObjects v1 with delimiter - objects: {:?}", contents);
-    assert_eq!(contents.len(), 1);
-    assert!(contents.contains(&"README.md"));
-
-    // Should have one common prefix: "dir/"
-    let prefixes: Vec<_> = response.common_prefixes().iter().filter_map(|cp| cp.prefix()).collect();
-    debug!("ListObjects v1 with delimiter - prefixes: {:?}", prefixes);
-    assert_eq!(prefixes.len(), 1);
-    assert!(prefixes.contains(&"dir/"));
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_list_objects_v1_next_marker_with_delimiter() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-v1-next-marker-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    // Objects: a.txt, dir/1.txt (→ common prefix "dir/"), z.txt
-    // With delimiter="/", max_keys=2 → page 1 = [a.txt, dir/], truncated
-    let keys = ["a.txt", "dir/1.txt", "z.txt"];
-    for key in &keys {
-        c.put_object()
-            .bucket(bucket)
-            .key(*key)
-            .body(ByteStream::from_static(b"x"))
-            .send()
-            .await?;
-    }
-
-    let page1 = c.list_objects().bucket(bucket).delimiter("/").max_keys(2).send().await?;
-
-    assert_eq!(page1.is_truncated(), Some(true));
-    // Per S3 spec: when delimiter is set and is_truncated is true,
-    // next_marker must be returned so the client can paginate.
-    assert!(
-        page1.next_marker().is_some(),
-        "is_truncated is true with delimiter but next_marker is missing"
-    );
-    let page1_contents: Vec<_> = page1.contents().iter().filter_map(|obj| obj.key()).collect();
-    let page1_prefixes: Vec<_> = page1.common_prefixes().iter().filter_map(|cp| cp.prefix()).collect();
-    assert_eq!(page1_contents, vec!["a.txt"]);
-    assert_eq!(page1_prefixes, vec!["dir/"]);
-    let next_marker = page1.next_marker().expect("truncated v1 listing should include next_marker");
-
-    let page2 = c
-        .list_objects()
-        .bucket(bucket)
-        .delimiter("/")
-        .max_keys(2)
-        .marker(next_marker)
-        .send()
-        .await?;
-
-    assert_eq!(page2.is_truncated(), Some(false));
-    let page2_contents: Vec<_> = page2.contents().iter().filter_map(|obj| obj.key()).collect();
-    let page2_prefixes: Vec<_> = page2.common_prefixes().iter().filter_map(|cp| cp.prefix()).collect();
-    assert_eq!(page2_contents, vec!["z.txt"]);
-    assert!(
-        page2_prefixes.is_empty(),
-        "second page should not repeat the already-returned common prefix"
-    );
-
-    // Cleanup
-    for key in &keys {
-        delete_object(&c, bucket, key).await?;
-    }
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// Walk all v1 pages using `marker` (no delimiter) and verify the full key set.
-#[tokio::test]
-#[tracing::instrument]
-async fn test_list_objects_v1_marker_pagination() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-v1-marker-pag-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    let keys = ["a.txt", "b.txt", "c.txt", "d.txt", "e.txt"];
-    for key in &keys {
-        c.put_object()
-            .bucket(bucket)
-            .key(*key)
-            .body(ByteStream::from_static(b"x"))
-            .send()
-            .await?;
-    }
-
-    // Walk all pages with max_keys=2
-    let mut all_keys: Vec<String> = Vec::new();
-    let mut marker: Option<String> = None;
-    loop {
-        let mut req = c.list_objects().bucket(bucket).max_keys(2);
-        if let Some(m) = &marker {
-            req = req.marker(m.clone());
-        }
-        let page = req.send().await?;
-
-        all_keys.extend(page.contents().iter().filter_map(|o| o.key().map(String::from)));
-
-        if page.is_truncated() != Some(true) {
-            break;
-        }
-        marker = page.next_marker().map(String::from);
-        assert!(marker.is_some(), "is_truncated is true but next_marker is missing");
-    }
-
-    assert_eq!(all_keys, vec!["a.txt", "b.txt", "c.txt", "d.txt", "e.txt"]);
-
-    for key in &keys {
-        delete_object(&c, bucket, key).await?;
-    }
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// Walk all v1 pages with `max_keys=1` to verify single-item pages advance correctly.
-#[tokio::test]
-#[tracing::instrument]
-async fn test_list_objects_v1_max_keys_one() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-v1-maxkeys1-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    let keys = ["a.txt", "b.txt", "c.txt"];
-    for key in &keys {
-        c.put_object()
-            .bucket(bucket)
-            .key(*key)
-            .body(ByteStream::from_static(b"x"))
-            .send()
-            .await?;
-    }
-
-    let mut all_keys: Vec<String> = Vec::new();
-    let mut marker: Option<String> = None;
-    let mut page_count = 0;
-    loop {
-        let mut req = c.list_objects().bucket(bucket).max_keys(1);
-        if let Some(m) = &marker {
-            req = req.marker(m.clone());
-        }
-        let page = req.send().await?;
-
-        let page_keys: Vec<_> = page.contents().iter().filter_map(|o| o.key().map(String::from)).collect();
-        assert!(page_keys.len() <= 1, "max_keys=1 but got {} results", page_keys.len());
-        all_keys.extend(page_keys);
-        page_count += 1;
-
-        if page.is_truncated() != Some(true) {
-            break;
-        }
-        marker = page.next_marker().map(String::from);
-        assert!(marker.is_some(), "is_truncated is true but next_marker is missing");
-    }
-
-    assert_eq!(all_keys, vec!["a.txt", "b.txt", "c.txt"]);
-    assert_eq!(page_count, 3, "expected 3 single-item pages");
-
-    for key in &keys {
-        delete_object(&c, bucket, key).await?;
-    }
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// Walk all v1 pages with `delimiter` and multiple common prefixes,
-/// verifying no duplicates and complete coverage.
-#[tokio::test]
-#[tracing::instrument]
-async fn test_list_objects_v1_delimiter_multi_prefix_pagination() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-v1-multi-pfx-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    // Objects sorted: a.txt, dir1/x.txt (→ "dir1/"), dir2/y.txt (→ "dir2/"), z.txt
-    // With delimiter="/", this produces entries: a.txt, dir1/, dir2/, z.txt
-    let keys = ["a.txt", "dir1/x.txt", "dir2/y.txt", "z.txt"];
-    for key in &keys {
-        c.put_object()
-            .bucket(bucket)
-            .key(*key)
-            .body(ByteStream::from_static(b"x"))
-            .send()
-            .await?;
-    }
-
-    let mut all_objects: Vec<String> = Vec::new();
-    let mut all_prefixes: Vec<String> = Vec::new();
-    let mut marker: Option<String> = None;
-    loop {
-        let mut req = c.list_objects().bucket(bucket).delimiter("/").max_keys(2);
-        if let Some(m) = &marker {
-            req = req.marker(m.clone());
-        }
-        let page = req.send().await?;
-
-        all_objects.extend(page.contents().iter().filter_map(|o| o.key().map(String::from)));
-        all_prefixes.extend(page.common_prefixes().iter().filter_map(|p| p.prefix().map(String::from)));
-
-        if page.is_truncated() != Some(true) {
-            break;
-        }
-        marker = page.next_marker().map(String::from);
-        assert!(marker.is_some(), "is_truncated is true but next_marker is missing");
-    }
-
-    assert_eq!(all_objects, vec!["a.txt", "z.txt"]);
-    assert_eq!(all_prefixes, vec!["dir1/", "dir2/"]);
-
-    for key in &keys {
-        delete_object(&c, bucket, key).await?;
-    }
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// Walk all v2 pages with `delimiter` and multiple common prefixes,
-/// verifying no duplicates and complete coverage.
-#[tokio::test]
-#[tracing::instrument]
-async fn test_list_objects_v2_delimiter_multi_prefix_pagination() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-v2-multi-pfx-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    let keys = ["a.txt", "dir1/x.txt", "dir2/y.txt", "z.txt"];
-    for key in &keys {
-        c.put_object()
-            .bucket(bucket)
-            .key(*key)
-            .body(ByteStream::from_static(b"x"))
-            .send()
-            .await?;
-    }
-
-    let mut all_objects: Vec<String> = Vec::new();
-    let mut all_prefixes: Vec<String> = Vec::new();
-    let mut token: Option<String> = None;
-    loop {
-        let mut req = c.list_objects_v2().bucket(bucket).delimiter("/").max_keys(2);
-        if let Some(t) = &token {
-            req = req.continuation_token(t.clone());
-        }
-        let page = req.send().await?;
-
-        all_objects.extend(page.contents().iter().filter_map(|o| o.key().map(String::from)));
-        all_prefixes.extend(page.common_prefixes().iter().filter_map(|p| p.prefix().map(String::from)));
-
-        if page.is_truncated() != Some(true) {
-            break;
-        }
-        token = page.next_continuation_token().map(String::from);
-        assert!(token.is_some(), "is_truncated is true but next_continuation_token is missing");
-    }
-
-    assert_eq!(all_objects, vec!["a.txt", "z.txt"]);
-    assert_eq!(all_prefixes, vec!["dir1/", "dir2/"]);
-
-    for key in &keys {
-        delete_object(&c, bucket, key).await?;
-    }
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_list_objects_v2_max_keys() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-max-keys-{}", Uuid::new_v4());
-    let bucket_str = bucket.as_str();
-    create_bucket(&c, bucket_str).await?;
-
-    // Create 10 files
-    let content = "test";
-    for i in 0..10 {
-        let key = format!("file{i:02}.txt");
-        c.put_object()
-            .bucket(bucket_str)
-            .key(key)
-            .body(ByteStream::from_static(content.as_bytes()))
-            .send()
-            .await?;
-    }
-
-    // Test max_keys=5
-    let result = c.list_objects_v2().bucket(bucket_str).max_keys(5).send().await;
-    let response = log_and_unwrap!(result);
-
-    // Should return exactly 5 objects
-    let contents: Vec<_> = response.contents().iter().filter_map(|obj| obj.key()).collect();
-    assert_eq!(contents.len(), 5, "Expected 5 objects, got {}", contents.len());
-    assert_eq!(response.key_count(), Some(5));
-    assert_eq!(response.max_keys(), Some(5));
-    assert_eq!(response.is_truncated(), Some(true), "Should be truncated");
-
-    // Test max_keys=20 (more than available)
-    let result = c.list_objects_v2().bucket(bucket_str).max_keys(20).send().await;
-    let response = log_and_unwrap!(result);
-
-    let contents: Vec<_> = response.contents().iter().filter_map(|obj| obj.key()).collect();
-    assert_eq!(contents.len(), 10, "Expected 10 objects, got {}", contents.len());
-    assert_eq!(response.key_count(), Some(10));
-    assert_eq!(response.max_keys(), Some(20));
-    assert_eq!(response.is_truncated(), Some(false), "Should not be truncated");
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_single_object() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-single-object-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    let key = "sample.txt";
-    let content = "hello world\n你好世界\n";
-    let crc32c = base64_simd::STANDARD.encode_to_string(crc32c::crc32c(content.as_bytes()).to_be_bytes());
-
-    create_bucket(&c, bucket).await?;
-
-    {
-        let body = ByteStream::from_static(content.as_bytes());
-        c.put_object()
-            .bucket(bucket)
-            .key(key)
-            .body(body)
-            .checksum_crc32_c(crc32c.as_str())
-            .send()
-            .await?;
-    }
-
-    {
-        let ans = c
-            .get_object()
-            .bucket(bucket)
-            .key(key)
-            .checksum_mode(ChecksumMode::Enabled)
-            .send()
-            .await?;
-
-        let content_length: usize = ans.content_length().unwrap().try_into().unwrap();
-        let checksum_crc32c = ans.checksum_crc32_c.unwrap();
-        let body = ans.body.collect().await?.into_bytes();
-
-        assert_eq!(content_length, content.len());
-        assert_eq!(checksum_crc32c, crc32c);
-        assert_eq!(body.as_ref(), content.as_bytes());
-    }
-
-    {
-        delete_object(&c, bucket, key).await?;
-        delete_bucket(&c, bucket).await?;
-    }
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_multipart() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-
-    let bucket = format!("test-multipart-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    let key = "sample.txt";
-    let content = "abcdefghijklmnopqrstuvwxyz/0123456789/!@#$%^&*();\n";
-
-    let upload_id = {
-        let ans = c.create_multipart_upload().bucket(bucket).key(key).send().await?;
-        ans.upload_id.unwrap()
-    };
-    let upload_id = upload_id.as_str();
-
-    let upload_parts = {
-        let body = ByteStream::from_static(content.as_bytes());
-        let part_number = 1;
-
-        let ans = c
-            .upload_part()
-            .bucket(bucket)
-            .key(key)
-            .upload_id(upload_id)
-            .body(body)
-            .part_number(part_number)
-            .send()
-            .await?;
-
-        let part = CompletedPart::builder()
-            .e_tag(ans.e_tag.unwrap_or_default())
-            .part_number(part_number)
-            .build();
-
-        vec![part]
-    };
-
-    {
-        let upload = CompletedMultipartUpload::builder().set_parts(Some(upload_parts)).build();
-
-        let _ = c
-            .complete_multipart_upload()
-            .bucket(bucket)
-            .key(key)
-            .multipart_upload(upload)
-            .upload_id(upload_id)
-            .send()
-            .await?;
-    }
-
-    {
-        let ans = c.get_object().bucket(bucket).key(key).send().await?;
-
-        let content_length: usize = ans.content_length().unwrap().try_into().unwrap();
-        let body = ans.body.collect().await?.into_bytes();
-
-        assert_eq!(content_length, content.len());
-        assert_eq!(body.as_ref(), content.as_bytes());
-    }
-
-    {
-        delete_object(&c, bucket, key).await?;
-        delete_bucket(&c, bucket).await?;
-    }
-
-    Ok(())
-}
-
-#[derive(Debug, Clone, Copy)]
-enum MultipartXxHashCase {
-    XxHash64,
-    XxHash3,
-    XxHash128,
-}
-
-impl MultipartXxHashCase {
-    fn name(self) -> &'static str {
-        match self {
-            Self::XxHash64 => "xxhash64",
-            Self::XxHash3 => "xxhash3",
-            Self::XxHash128 => "xxhash128",
-        }
-    }
-
-    fn algorithm(self) -> aws_sdk_s3::types::ChecksumAlgorithm {
-        use aws_sdk_s3::types::ChecksumAlgorithm;
-
-        match self {
-            Self::XxHash64 => ChecksumAlgorithm::Xxhash64,
-            Self::XxHash3 => ChecksumAlgorithm::Xxhash3,
-            Self::XxHash128 => ChecksumAlgorithm::Xxhash128,
-        }
-    }
-
-    fn expected_checksum(self, content: &[u8]) -> String {
-        use s3s::crypto::Checksum as _;
-        use s3s::crypto::XxHash3;
-        use s3s::crypto::XxHash64;
-        use s3s::crypto::XxHash128;
-
-        match self {
-            Self::XxHash64 => base64_simd::STANDARD.encode_to_string(XxHash64::checksum(content)),
-            Self::XxHash3 => base64_simd::STANDARD.encode_to_string(XxHash3::checksum(content)),
-            Self::XxHash128 => base64_simd::STANDARD.encode_to_string(XxHash128::checksum(content)),
-        }
-    }
-
-    fn upload_part_checksum(self, output: &aws_sdk_s3::operation::upload_part::UploadPartOutput) -> Option<&str> {
-        match self {
-            Self::XxHash64 => output.checksum_xxhash64(),
-            Self::XxHash3 => output.checksum_xxhash3(),
-            Self::XxHash128 => output.checksum_xxhash128(),
-        }
-    }
-
-    fn listed_part_checksum(self, part: &aws_sdk_s3::types::Part) -> Option<&str> {
-        match self {
-            Self::XxHash64 => part.checksum_xxhash64(),
-            Self::XxHash3 => part.checksum_xxhash3(),
-            Self::XxHash128 => part.checksum_xxhash128(),
-        }
-    }
-
-    fn completed_part(self, e_tag: String, checksum: String, part_number: i32) -> CompletedPart {
-        match self {
-            Self::XxHash64 => CompletedPart::builder()
-                .e_tag(e_tag)
-                .checksum_xxhash64(checksum)
-                .part_number(part_number)
-                .build(),
-            Self::XxHash3 => CompletedPart::builder()
-                .e_tag(e_tag)
-                .checksum_xxhash3(checksum)
-                .part_number(part_number)
-                .build(),
-            Self::XxHash128 => CompletedPart::builder()
-                .e_tag(e_tag)
-                .checksum_xxhash128(checksum)
-                .part_number(part_number)
-                .build(),
-        }
-    }
-
-    fn complete_checksum(
-        self,
-        output: &aws_sdk_s3::operation::complete_multipart_upload::CompleteMultipartUploadOutput,
-    ) -> Option<&str> {
-        match self {
-            Self::XxHash64 => output.checksum_xxhash64(),
-            Self::XxHash3 => output.checksum_xxhash3(),
-            Self::XxHash128 => output.checksum_xxhash128(),
-        }
-    }
-
-    fn head_checksum(self, output: &aws_sdk_s3::operation::head_object::HeadObjectOutput) -> Option<&str> {
-        match self {
-            Self::XxHash64 => output.checksum_xxhash64(),
-            Self::XxHash3 => output.checksum_xxhash3(),
-            Self::XxHash128 => output.checksum_xxhash128(),
-        }
-    }
-
-    fn get_checksum(self, output: &aws_sdk_s3::operation::get_object::GetObjectOutput) -> Option<&str> {
-        match self {
-            Self::XxHash64 => output.checksum_xxhash64(),
-            Self::XxHash3 => output.checksum_xxhash3(),
-            Self::XxHash128 => output.checksum_xxhash128(),
-        }
-    }
-}
-
-async fn assert_multipart_xxhash_checksums(client: &Client, bucket: &str, checksum_case: MultipartXxHashCase) -> Result<()> {
-    let key = format!("multipart-{}.txt", checksum_case.name());
-    let content = format!("multipart checksum coverage via {}\n", checksum_case.name());
-    let expected_checksum = checksum_case.expected_checksum(content.as_bytes());
-
-    let upload_id = client
-        .create_multipart_upload()
-        .bucket(bucket)
-        .key(key.as_str())
-        .checksum_algorithm(checksum_case.algorithm())
-        .send()
-        .await?
-        .upload_id
-        .expect("create_multipart_upload should return upload_id");
-
-    let part_number = 1;
-    let upload_part = client
-        .upload_part()
-        .bucket(bucket)
-        .key(key.as_str())
-        .upload_id(upload_id.as_str())
-        .body(ByteStream::from(content.into_bytes()))
-        .part_number(part_number)
-        .send()
-        .await?;
-
-    let part_etag = upload_part.e_tag().expect("upload_part should return e_tag").to_owned();
-    let part_checksum = checksum_case
-        .upload_part_checksum(&upload_part)
-        .unwrap_or_else(|| panic!("upload_part should return checksum_{}", checksum_case.name()))
-        .to_owned();
-    assert_eq!(part_checksum, expected_checksum, "upload_part checksum mismatch for {checksum_case:?}");
-
-    let listed = client
-        .list_parts()
-        .bucket(bucket)
-        .key(key.as_str())
-        .upload_id(upload_id.as_str())
-        .send()
-        .await?;
-    let listed_part = listed.parts().first().expect("list_parts should return one part");
-    let listed_checksum = checksum_case.listed_part_checksum(listed_part);
-    assert_eq!(
-        listed_checksum,
-        Some(expected_checksum.as_str()),
-        "list_parts checksum mismatch for {checksum_case:?}"
-    );
-
-    let completed_part = checksum_case.completed_part(part_etag, part_checksum, part_number);
-    let completed_upload = CompletedMultipartUpload::builder()
-        .set_parts(Some(vec![completed_part]))
-        .build();
-
-    let complete = client
-        .complete_multipart_upload()
-        .bucket(bucket)
-        .key(key.as_str())
-        .multipart_upload(completed_upload)
-        .upload_id(upload_id.as_str())
-        .send()
-        .await?;
-    let complete_checksum = checksum_case.complete_checksum(&complete);
-    assert_eq!(
-        complete_checksum,
-        Some(expected_checksum.as_str()),
-        "complete_multipart_upload checksum mismatch for {checksum_case:?}"
-    );
-    assert_eq!(complete.checksum_type().map(aws_sdk_s3::types::ChecksumType::as_str), Some("FULL_OBJECT"));
-
-    let head = client.head_object().bucket(bucket).key(key.as_str()).send().await?;
-    let head_checksum = checksum_case.head_checksum(&head);
-    assert_eq!(
-        head_checksum,
-        Some(expected_checksum.as_str()),
-        "head_object checksum mismatch for {checksum_case:?}"
-    );
-
-    let get = client
-        .get_object()
-        .bucket(bucket)
-        .key(key.as_str())
-        .checksum_mode(ChecksumMode::Enabled)
-        .send()
-        .await?;
-    let get_checksum = checksum_case.get_checksum(&get);
-    assert_eq!(
-        get_checksum,
-        Some(expected_checksum.as_str()),
-        "get_object checksum mismatch for {checksum_case:?}"
-    );
-
-    delete_object(client, bucket, key.as_str()).await?;
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_multipart_xxhash_checksums() -> Result<()> {
-    let _guard = serial().await;
-
-    let client = Client::new(config());
-
-    let bucket = format!("test-multipart-xxhash-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&client, bucket).await?;
-
-    for checksum_case in [
-        MultipartXxHashCase::XxHash64,
-        MultipartXxHashCase::XxHash3,
-        MultipartXxHashCase::XxHash128,
-    ] {
-        assert_multipart_xxhash_checksums(&client, bucket, checksum_case).await?;
-    }
-
-    delete_bucket(&client, bucket).await?;
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_multipart_checksum_type_composite_not_implemented() -> Result<()> {
-    use aws_sdk_s3::types::ChecksumAlgorithm;
-    use aws_sdk_s3::types::ChecksumType;
-
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-multipart-composite-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    let err = c
-        .create_multipart_upload()
-        .bucket(bucket)
-        .key("composite.txt")
-        .checksum_algorithm(ChecksumAlgorithm::Xxhash64)
-        .checksum_type(ChecksumType::Composite)
-        .send()
-        .await
-        .expect_err("COMPOSITE checksum_type should be rejected until implemented");
-    let service_err = err.into_service_error();
-    assert_eq!(service_err.code(), Some("NotImplemented"));
-
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// Test that multipart uploaded objects have the correct `ETag` format: `{hash}-{part_count}`
-#[tokio::test]
-#[tracing::instrument]
-async fn test_multipart_etag_format() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-
-    let bucket = format!("test-multipart-etag-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    let key = "multipart-etag.txt";
-    let content = "abcdefghijklmnopqrstuvwxyz/0123456789/!@#$%^&*();\n";
-
-    let upload_id = {
-        let ans = c.create_multipart_upload().bucket(bucket).key(key).send().await?;
-        ans.upload_id.unwrap()
-    };
-    let upload_id = upload_id.as_str();
-
-    let upload_parts = {
-        let body = ByteStream::from_static(content.as_bytes());
-        let part_number = 1;
-
-        let ans = c
-            .upload_part()
-            .bucket(bucket)
-            .key(key)
-            .upload_id(upload_id)
-            .body(body)
-            .part_number(part_number)
-            .send()
-            .await?;
-
-        let part = CompletedPart::builder()
-            .e_tag(ans.e_tag.expect("upload_part response missing e_tag"))
-            .part_number(part_number)
-            .build();
-
-        vec![part]
-    };
-
-    let complete_e_tag = {
-        let upload = CompletedMultipartUpload::builder().set_parts(Some(upload_parts)).build();
-
-        let ans = c
-            .complete_multipart_upload()
-            .bucket(bucket)
-            .key(key)
-            .multipart_upload(upload)
-            .upload_id(upload_id)
-            .send()
-            .await?;
-
-        let e_tag = ans.e_tag().unwrap().to_owned();
-        debug!(?e_tag, "multipart etag");
-
-        // Multipart ETags must have the format: {hex_md5}-{part_count}
-        let unquoted = e_tag.trim_matches('"');
-        let (hash_part, count_part) = unquoted.rsplit_once('-').expect("multipart ETag should contain a dash");
-        assert_eq!(hash_part.len(), 32, "hash part should be 32 hex characters: {hash_part}");
-        assert!(hash_part.chars().all(|c| c.is_ascii_hexdigit()), "hash part should be hex: {hash_part}");
-        let part_count: usize = count_part.parse().expect("count part should be a number");
-        assert_eq!(part_count, 1, "part count should match number of parts uploaded");
-
-        e_tag
-    };
-
-    {
-        // Verify the ETag from head_object matches complete_multipart_upload
-        let ans = c.head_object().bucket(bucket).key(key).send().await?;
-        let head_e_tag = ans.e_tag().unwrap();
-        debug!(?head_e_tag, "head_object etag");
-        assert_eq!(head_e_tag, complete_e_tag, "head_object ETag should match complete_multipart_upload ETag");
-    }
-
-    {
-        // Verify the ETag from get_object matches complete_multipart_upload
-        let ans = c.get_object().bucket(bucket).key(key).send().await?;
-        let get_e_tag = ans.e_tag().unwrap();
-        debug!(?get_e_tag, "get_object etag");
-        assert_eq!(get_e_tag, complete_e_tag, "get_object ETag should match complete_multipart_upload ETag");
-    }
-
-    {
-        delete_object(&c, bucket, key).await?;
-        delete_bucket(&c, bucket).await?;
-    }
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_upload_part_copy() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let src_bucket = format!("test-copy{}", Uuid::new_v4());
-    let src_bucket = src_bucket.as_str();
-    let src_key = "copied.txt";
-    let src_content = "hello world\nनमस्ते दुनिया\n";
-    let crc32c = base64_simd::STANDARD.encode_to_string(crc32c::crc32c(src_content.as_bytes()).to_be_bytes());
-
-    create_bucket(&c, src_bucket).await?;
-
-    {
-        let src_body = ByteStream::from_static(src_content.as_bytes());
-        c.put_object()
-            .bucket(src_bucket)
-            .key(src_key)
-            .body(src_body)
-            .checksum_crc32_c(crc32c.as_str())
-            .send()
-            .await?;
-    }
-
-    let bucket = format!("test-uploadpartcopy-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    let key = "sample.txt";
-
-    let upload_id = {
-        use aws_sdk_s3::types::ChecksumAlgorithm;
-
-        let ans = c
-            .create_multipart_upload()
-            .bucket(bucket)
-            .key(key)
-            .checksum_algorithm(ChecksumAlgorithm::Crc32C)
-            .send()
-            .await?;
-        ans.upload_id.unwrap()
-    };
-    let upload_id = upload_id.as_str();
-    let src_path = format!("{src_bucket}/{src_key}");
-    let upload_parts = {
-        let part_number = 1;
-        let ans = c
-            .upload_part_copy()
-            .bucket(bucket)
-            .key(key)
-            .copy_source(src_path)
-            .upload_id(upload_id)
-            .part_number(part_number)
-            .send()
-            .await?;
-
-        let copy_part_result = ans
-            .copy_part_result()
-            .expect("upload_part_copy should return copy_part_result");
-        let copied_checksum_crc32c = copy_part_result
-            .checksum_crc32_c()
-            .expect("upload_part_copy should return checksum_crc32_c")
-            .to_owned();
-        assert_eq!(copied_checksum_crc32c, crc32c);
-
-        let part = CompletedPart::builder()
-            .checksum_crc32_c(copied_checksum_crc32c)
-            .e_tag(copy_part_result.e_tag().expect("copy_part_result should return e_tag"))
-            .part_number(part_number)
-            .build();
-        vec![part]
-    };
-
-    {
-        let upload = CompletedMultipartUpload::builder().set_parts(Some(upload_parts)).build();
-
-        let _ = c
-            .complete_multipart_upload()
-            .bucket(bucket)
-            .key(key)
-            .multipart_upload(upload)
-            .upload_id(upload_id)
-            .send()
-            .await?;
-    }
-
-    {
-        let ans = c
-            .get_object()
-            .bucket(bucket)
-            .key(key)
-            .checksum_mode(ChecksumMode::Enabled)
-            .send()
-            .await?;
-        assert_eq!(ans.checksum_crc32_c(), Some(crc32c.as_str()));
-
-        let content_length: usize = ans.content_length().unwrap().try_into().unwrap();
-        let body = ans.body.collect().await?.into_bytes();
-
-        assert_eq!(content_length, src_content.len());
-        assert_eq!(body.as_ref(), src_content.as_bytes());
-    }
-
-    {
-        delete_object(&c, bucket, key).await?;
-        delete_bucket(&c, bucket).await?;
-        delete_object(&c, src_bucket, src_key).await?;
-        delete_bucket(&c, src_bucket).await?;
-    }
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_upload_part_copy_invalid_source_range() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-upc-bad-range-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    let src_key = "src.txt";
-    let src_content = "hello";
-    c.put_object()
-        .bucket(bucket)
-        .key(src_key)
-        .body(ByteStream::from_static(src_content.as_bytes()))
-        .send()
-        .await?;
-
-    let dst_key = "dst.txt";
-    let upload_id = c
-        .create_multipart_upload()
-        .bucket(bucket)
-        .key(dst_key)
-        .send()
-        .await?
-        .upload_id
-        .expect("upload_id");
-    let upload_id = upload_id.as_str();
-
-    let copy_source = format!("{bucket}/{src_key}");
-    // Object length is 5; inclusive end index must be <= 4. End 5 is past EOF and must not truncate.
-    let err = c
-        .upload_part_copy()
-        .bucket(bucket)
-        .key(dst_key)
-        .copy_source(&copy_source)
-        .copy_source_range("bytes=0-5")
-        .upload_id(upload_id)
-        .part_number(1)
-        .send()
-        .await
-        .expect_err("Expected InvalidRange when copy range end is past EOF");
-    let service_err = err.into_service_error();
-    assert_eq!(
-        service_err.code(),
-        Some("InvalidRange"),
-        "past-EOF range: expected InvalidRange, got {:?}",
-        service_err.code()
-    );
-
-    let err = c
-        .upload_part_copy()
-        .bucket(bucket)
-        .key(dst_key)
-        .copy_source(&copy_source)
-        .copy_source_range("bytes=0-18446744073709551615")
-        .upload_id(upload_id)
-        .part_number(1)
-        .send()
-        .await
-        .expect_err("Expected InvalidRange for end=u64::MAX");
-    let service_err = err.into_service_error();
-    assert_eq!(
-        service_err.code(),
-        Some("InvalidRange"),
-        "u64::MAX end: expected InvalidRange, got {:?}",
-        service_err.code()
-    );
-
-    c.abort_multipart_upload()
-        .bucket(bucket)
-        .key(dst_key)
-        .upload_id(upload_id)
-        .send()
-        .await?;
-
-    delete_object(&c, bucket, src_key).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_single_object_get_range() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-single-object-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    let key = "sample.txt";
-    let content = "hello world\n你好世界\n";
-    let crc32c = base64_simd::STANDARD.encode_to_string(crc32c::crc32c(content.as_bytes()).to_be_bytes());
-
-    create_bucket(&c, bucket).await?;
-
-    {
-        let body = ByteStream::from_static(content.as_bytes());
-        c.put_object()
-            .bucket(bucket)
-            .key(key)
-            .body(body)
-            .checksum_crc32_c(crc32c.as_str())
-            .send()
-            .await?;
-    }
-
-    {
-        let ans = c
-            .get_object()
-            .bucket(bucket)
-            .key(key)
-            .range("bytes=0-4")
-            .checksum_mode(ChecksumMode::Enabled)
-            .send()
-            .await?;
-
-        // S3 doesn't return checksums when a range is specified
-        assert!(&ans.checksum_crc32().is_none());
-        assert!(&ans.checksum_crc32_c().is_none());
-
-        let content_length: usize = ans.content_length().unwrap().try_into().unwrap();
-        let body = ans.body.collect().await?.into_bytes();
-
-        assert_eq!(content_length, 5);
-        assert_eq!(body.as_ref(), &content.as_bytes()[0..=4]);
-    }
-
-    {
-        let ans = c
-            .get_object()
-            .bucket(bucket)
-            .key(key)
-            .range("bytes=0-1000")
-            .checksum_mode(ChecksumMode::Enabled)
-            .send()
-            .await?;
-
-        let content_length: usize = ans.content_length().unwrap().try_into().unwrap();
-        let checksum_crc32c = ans.checksum_crc32_c.unwrap();
-        let body = ans.body.collect().await?.into_bytes();
-
-        assert_eq!(content_length, content.len());
-        assert_eq!(checksum_crc32c, crc32c);
-        assert_eq!(body.as_ref(), content.as_bytes());
-    }
-
-    {
-        delete_object(&c, bucket, key).await?;
-        delete_bucket(&c, bucket).await?;
-    }
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_relaxed_bucket_validation() -> Result<()> {
-    struct RelaxedNameValidation;
-
-    impl NameValidation for RelaxedNameValidation {
-        fn validate_bucket_name(&self, name: &str) -> bool {
-            !name.is_empty()
-        }
-    }
-
-    let _guard = serial().await;
-
-    let c = create_client_with_validation(RelaxedNameValidation);
-
-    // Test with bucket names that should pass with relaxed validation
-    let relaxed_bucket_names = [
-        "UPPERCASE-BUCKET",       // Uppercase not normally allowed
-        "bucket_with_underscore", // Underscores not allowed
-    ];
-
-    for bucket_name in relaxed_bucket_names {
-        let location = BucketLocationConstraint::from(REGION);
-        let cfg = CreateBucketConfiguration::builder().location_constraint(location).build();
-
-        let result = c
-            .create_bucket()
-            .create_bucket_configuration(cfg)
-            .bucket(bucket_name)
-            .send()
-            .await;
-
-        // Should not fail due to bucket name validation
-        match result {
-            Ok(_) => {
-                debug!("Successfully created bucket with relaxed validation: {bucket_name}");
-
-                // Verify the bucket was actually created by checking bucket existence
-                let head_result = c.head_bucket().bucket(bucket_name).send().await;
-                assert!(head_result.is_ok(), "Failed to head bucket {bucket_name} after creation");
-
-                // Clean up the bucket
-                let delete_result = delete_bucket(&c, bucket_name).await;
-                assert!(delete_result.is_ok(), "Failed to delete bucket {bucket_name}");
-            }
-            Err(e) => {
-                let error_str = format!("{e:?}");
-                debug!("Bucket creation failed for other reasons (expected): {bucket_name} - {error_str}");
-                // Verify it's not a bucket name validation error
-                assert!(!error_str.contains("InvalidBucketName") && !error_str.contains("bucket name"));
-            }
-        }
-    }
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_default_bucket_validation() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config()); // Uses default validation
-
-    // Test with invalid bucket names that should be rejected by AWS rules
-    let invalid_bucket_names = [
-        "UPPERCASE-BUCKET",       // Uppercase not allowed
-        "bucket_with_underscore", // Underscores not allowed
-        "bucket..double.dots",    // Consecutive dots not allowed
-    ];
-
-    for bucket_name in invalid_bucket_names {
-        // Try to create bucket with invalid name - should fail with default validation
-        let location = BucketLocationConstraint::from(REGION);
-        let cfg = CreateBucketConfiguration::builder().location_constraint(location).build();
-
-        let result = c
-            .create_bucket()
-            .create_bucket_configuration(cfg)
-            .bucket(bucket_name)
-            .send()
-            .await;
-
-        // Should fail due to bucket name validation
-        assert!(result.is_err(), "Expected error for invalid bucket name: {bucket_name}");
-
-        let error_str = format!("{:?}", result.unwrap_err());
-        debug!("Default validation rejected bucket name {bucket_name}: {error_str}");
-    }
-
-    Ok(())
-}
-
-/// Test that demonstrates the Content-Encoding preservation issue
-/// Related: <https://github.com/rustfs/rustfs/issues/1062>
-#[tokio::test]
-#[tracing::instrument]
-async fn test_content_encoding_preservation() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-content-encoding-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    let key = "compressed.json";
-
-    // Simulated Brotli-compressed JSON content
-    let content = b"compressed data here";
-
-    create_bucket(&c, bucket).await?;
-
-    // Upload object with Content-Encoding header
-    {
-        let body = ByteStream::from_static(content);
-        c.put_object()
-            .bucket(bucket)
-            .key(key)
-            .body(body)
-            .content_encoding("br") // Brotli compression
-            .content_type("application/json")
-            .content_disposition("attachment; filename=\"data.json\"")
-            .cache_control("max-age=3600")
-            .send()
-            .await?;
-
-        debug!("Uploaded object with Content-Encoding: br");
-    }
-
-    // Retrieve object and verify headers are preserved
-    {
-        let ans = c.get_object().bucket(bucket).key(key).send().await?;
-
-        // Verify that standard object attributes are now preserved by s3s-fs
-        debug!("Retrieved object:");
-        debug!("  Content-Encoding: {:?}", ans.content_encoding());
-        debug!("  Content-Type: {:?}", ans.content_type());
-        debug!("  Content-Disposition: {:?}", ans.content_disposition());
-        debug!("  Cache-Control: {:?}", ans.cache_control());
-
-        // All standard attributes should be preserved
-        assert_eq!(ans.content_encoding(), Some("br"));
-        assert_eq!(ans.content_type(), Some("application/json"));
-        assert_eq!(ans.content_disposition(), Some("attachment; filename=\"data.json\""));
-        assert_eq!(ans.cache_control(), Some("max-age=3600"));
-    }
-
-    // Also test HeadObject
-    {
-        let ans = c.head_object().bucket(bucket).key(key).send().await?;
-
-        debug!("HeadObject result:");
-        debug!("  Content-Encoding: {:?}", ans.content_encoding());
-        debug!("  Content-Type: {:?}", ans.content_type());
-
-        // Verify HeadObject also returns the stored attributes
-        assert_eq!(ans.content_encoding(), Some("br"));
-        assert_eq!(ans.content_type(), Some("application/json"));
-    }
-
-    {
-        delete_object(&c, bucket, key).await?;
-        delete_bucket(&c, bucket).await?;
-    }
-
-    Ok(())
-}
-
-/// Test that standard object attributes are preserved through multipart uploads
-#[tokio::test]
-#[tracing::instrument]
-async fn test_multipart_with_attributes() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-multipart-attrs-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    let key = "multipart-with-attrs.json";
-
-    create_bucket(&c, bucket).await?;
-
-    // Create multipart upload with standard attributes
-    let upload_id = {
-        let ans = c
-            .create_multipart_upload()
-            .bucket(bucket)
-            .key(key)
-            .content_encoding("gzip")
-            .content_type("application/json")
-            .content_disposition("attachment; filename=\"data.json\"")
-            .cache_control("public, max-age=7200")
-            .send()
-            .await?;
-        ans.upload_id.unwrap()
-    };
-    let upload_id = upload_id.as_str();
-
-    // Upload a part
-    let content = b"part1 content";
-    let upload_parts = {
-        let body = ByteStream::from_static(content);
-        let part_number = 1;
-
-        let ans = c
-            .upload_part()
-            .bucket(bucket)
-            .key(key)
-            .upload_id(upload_id)
-            .body(body)
-            .part_number(part_number)
-            .send()
-            .await?;
-
-        let part = CompletedPart::builder()
-            .e_tag(ans.e_tag.unwrap_or_default())
-            .part_number(part_number)
-            .build();
-
-        vec![part]
-    };
-
-    // Complete the multipart upload
-    {
-        let upload = CompletedMultipartUpload::builder().set_parts(Some(upload_parts)).build();
-
-        c.complete_multipart_upload()
-            .bucket(bucket)
-            .key(key)
-            .multipart_upload(upload)
-            .upload_id(upload_id)
-            .send()
-            .await?;
-    }
-
-    // Verify attributes were preserved after completing multipart upload
-    {
-        let ans = c.get_object().bucket(bucket).key(key).send().await?;
-
-        debug!("Retrieved multipart object:");
-        debug!("  Content-Encoding: {:?}", ans.content_encoding());
-        debug!("  Content-Type: {:?}", ans.content_type());
-        debug!("  Content-Disposition: {:?}", ans.content_disposition());
-        debug!("  Cache-Control: {:?}", ans.cache_control());
-
-        // Verify all attributes are preserved through multipart upload
-        assert_eq!(ans.content_encoding(), Some("gzip"));
-        assert_eq!(ans.content_type(), Some("application/json"));
-        assert_eq!(ans.content_disposition(), Some("attachment; filename=\"data.json\""));
-        assert_eq!(ans.cache_control(), Some("public, max-age=7200"));
-    }
-
-    // Also verify with HeadObject
-    {
-        let ans = c.head_object().bucket(bucket).key(key).send().await?;
-
-        assert_eq!(ans.content_encoding(), Some("gzip"));
-        assert_eq!(ans.content_type(), Some("application/json"));
-    }
-
-    {
-        delete_object(&c, bucket, key).await?;
-        delete_bucket(&c, bucket).await?;
-    }
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_sts_assume_role_not_implemented() -> Result<()> {
-    let _guard = serial().await;
-
-    // Create STS client using the same config as S3
-    let sdk_config = config();
-    let sts_client = aws_sdk_sts::Client::new(sdk_config);
-
-    // Attempt to call AssumeRole - should fail with NotImplemented
-    let result = sts_client
-        .assume_role()
-        .role_arn("arn:aws:iam::123456789012:role/test-role")
-        .role_session_name("test-session")
-        .send()
-        .await;
-
-    // Verify the operation returned an error
-    assert!(result.is_err(), "Expected AssumeRole to fail with NotImplemented error");
-
-    // Check that the error is NotImplemented
-    let error = result.unwrap_err();
-    let error_str = format!("{error:?}");
-    debug!("AssumeRole error (expected): {error_str}");
-
-    // The error should contain "NotImplemented" or similar indication
-    assert!(
-        error_str.contains("NotImplemented") || error_str.contains("not implemented"),
-        "Expected NotImplemented error, got: {error_str}"
-    );
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_if_none_match_wildcard() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("if-none-match-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    let key = "test-file.txt";
-    let content1 = "initial content";
-    let content2 = "updated content";
-
-    create_bucket(&c, bucket).await?;
-
-    // Test 1: PUT with If-None-Match: * should succeed when object doesn't exist
-    debug!("Test 1: PUT with If-None-Match: * on non-existent object");
-    {
-        let body = ByteStream::from_static(content1.as_bytes());
-        let result = c
-            .put_object()
-            .bucket(bucket)
-            .key(key)
-            .body(body)
-            .if_none_match("*")
-            .send()
-            .await;
-
-        match result {
-            Ok(_) => debug!("✓ Successfully created object with If-None-Match: *"),
-            Err(e) => panic!("Expected PUT with If-None-Match: * to succeed when object doesn't exist, but got error: {e:?}"),
-        }
-    }
-
-    // Verify the object was created
-    {
-        let result = c.get_object().bucket(bucket).key(key).send().await?;
-        let body = result.body.collect().await?.into_bytes();
-        assert_eq!(body.as_ref(), content1.as_bytes());
-        debug!("✓ Verified object was created");
-    }
-
-    // Test 2: PUT with If-None-Match: * should fail when object exists
-    debug!("Test 2: PUT with If-None-Match: * on existing object");
-    {
-        let body = ByteStream::from_static(content2.as_bytes());
-        let result = c
-            .put_object()
-            .bucket(bucket)
-            .key(key)
-            .body(body)
-            .if_none_match("*")
-            .send()
-            .await;
-
-        match result {
-            Ok(_) => panic!("Expected PUT with If-None-Match: * to fail when object exists, but it succeeded"),
-            Err(e) => {
-                let error_str = format!("{e:?}");
-                debug!("✓ Expected error when object exists: {error_str}");
-                // The error should be a PreconditionFailed (412)
-                assert!(
-                    error_str.contains("PreconditionFailed") || error_str.contains("412"),
-                    "Expected PreconditionFailed error, got: {error_str}"
-                );
-            }
-        }
-    }
-
-    // Verify the object wasn't overwritten
-    {
-        let result = c.get_object().bucket(bucket).key(key).send().await?;
-        let body = result.body.collect().await?.into_bytes();
-        assert_eq!(body.as_ref(), content1.as_bytes());
-        debug!("✓ Verified object was not overwritten");
-    }
-
-    // Cleanup
-    delete_object(&c, bucket, key).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// Test that `PutObject` with `If-Match: *` succeeds when the object exists
-/// and fails with `PreconditionFailed` (412) when the object is absent.
-#[tokio::test]
-#[tracing::instrument]
-async fn test_put_object_if_match_wildcard() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("if-match-wc-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    let key = "test-file.txt";
-    let initial = "initial content";
-    let updated = "updated content";
-
-    create_bucket(&c, bucket).await?;
-
-    // Test 1: PUT with If-Match: * should fail when the object doesn't exist
-    {
-        let body = ByteStream::from_static(initial.as_bytes());
-        let err = c
-            .put_object()
-            .bucket(bucket)
-            .key(key)
-            .body(body)
-            .if_match("*")
-            .send()
-            .await
-            .expect_err("Expected If-Match: * on absent object to fail");
-
-        let service_err = err.into_service_error();
-        assert_eq!(
-            service_err.code(),
-            Some("PreconditionFailed"),
-            "Expected PreconditionFailed, got: {:?}",
-            service_err.code()
-        );
-    }
-
-    // Seed the object so the next steps have something to match against.
-    c.put_object()
-        .bucket(bucket)
-        .key(key)
-        .body(ByteStream::from_static(initial.as_bytes()))
-        .send()
-        .await?;
-
-    // Test 2: PUT with If-Match: * should succeed when the object exists
-    {
-        let body = ByteStream::from_static(updated.as_bytes());
-        c.put_object()
-            .bucket(bucket)
-            .key(key)
-            .body(body)
-            .if_match("*")
-            .send()
-            .await
-            .expect("Expected If-Match: * on existing object to succeed");
-    }
-
-    {
-        let result = c.get_object().bucket(bucket).key(key).send().await?;
-        let body = result.body.collect().await?.into_bytes();
-        assert_eq!(body.as_ref(), updated.as_bytes());
-    }
-
-    delete_object(&c, bucket, key).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// Test that `PutObject` with `If-Match: <etag>` overwrites only when the
-/// stored `ETag` matches and returns `PreconditionFailed` (412) otherwise.
-#[tokio::test]
-#[tracing::instrument]
-async fn test_put_object_if_match_etag() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("if-match-etag-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    let key = "test-file.txt";
-    let initial = "initial content";
-    let updated = "updated content";
-
-    create_bucket(&c, bucket).await?;
-
-    // Test 1: PUT with If-Match: <etag> should fail when the object doesn't exist
-    {
-        let body = ByteStream::from_static(initial.as_bytes());
-        let err = c
-            .put_object()
-            .bucket(bucket)
-            .key(key)
-            .body(body)
-            .if_match("\"some-etag\"")
-            .send()
-            .await
-            .expect_err("Expected If-Match on absent object to fail");
-
-        let service_err = err.into_service_error();
-        assert_eq!(
-            service_err.code(),
-            Some("PreconditionFailed"),
-            "Expected PreconditionFailed, got: {:?}",
-            service_err.code()
-        );
-    }
-
-    // Seed the object and capture the real ETag.
-    let initial_etag = {
-        let body = ByteStream::from_static(initial.as_bytes());
-        let result = c.put_object().bucket(bucket).key(key).body(body).send().await?;
-        result.e_tag().expect("put_object should return e_tag").to_owned()
-    };
-
-    // Test 2: PUT with a wrong ETag should fail and not overwrite
-    {
-        let body = ByteStream::from_static(updated.as_bytes());
-        let err = c
-            .put_object()
-            .bucket(bucket)
-            .key(key)
-            .body(body)
-            .if_match("\"wrong-etag-value\"")
-            .send()
-            .await
-            .expect_err("Expected If-Match with wrong ETag to fail");
-
-        let service_err = err.into_service_error();
-        assert_eq!(
-            service_err.code(),
-            Some("PreconditionFailed"),
-            "Expected PreconditionFailed, got: {:?}",
-            service_err.code()
-        );
-
-        let result = c.get_object().bucket(bucket).key(key).send().await?;
-        let body = result.body.collect().await?.into_bytes();
-        assert_eq!(body.as_ref(), initial.as_bytes(), "Object should not be overwritten");
-    }
-
-    // Test 3: PUT with the matching ETag should succeed and replace the body
-    {
-        let body = ByteStream::from_static(updated.as_bytes());
-        c.put_object()
-            .bucket(bucket)
-            .key(key)
-            .body(body)
-            .if_match(&initial_etag)
-            .send()
-            .await
-            .expect("Expected If-Match with matching ETag to succeed");
-
-        let result = c.get_object().bucket(bucket).key(key).send().await?;
-        let body = result.body.collect().await?.into_bytes();
-        assert_eq!(body.as_ref(), updated.as_bytes());
-    }
-
-    delete_object(&c, bucket, key).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_put_object_if_match_multipart_etag() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("if-match-multipart-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    let key = "multipart-source.txt";
-    let initial = b"multipart initial content";
-    let updated = b"updated through put object";
-
-    create_bucket(&c, bucket).await?;
-
-    let multipart_etag = {
-        let (upload_id, upload_parts) = do_multipart_upload(&c, bucket, key, initial).await?;
-        let upload = CompletedMultipartUpload::builder().set_parts(Some(upload_parts)).build();
-        let result = c
-            .complete_multipart_upload()
-            .bucket(bucket)
-            .key(key)
-            .upload_id(&upload_id)
-            .multipart_upload(upload)
-            .send()
-            .await?;
-        result
-            .e_tag()
-            .expect("complete_multipart_upload should return e_tag")
-            .to_owned()
-    };
-    assert!(multipart_etag.contains('-'), "expected multipart ETag format");
-
-    let err = c
-        .put_object()
-        .bucket(bucket)
-        .key(key)
-        .body(ByteStream::from_static(b"wrong overwrite"))
-        .if_match("\"wrong-etag-value\"")
-        .send()
-        .await
-        .expect_err("wrong ETag should fail");
-    assert_eq!(err.into_service_error().code(), Some("PreconditionFailed"));
-
-    let result = c.get_object().bucket(bucket).key(key).send().await?;
-    let body = result.body.collect().await?.into_bytes();
-    assert_eq!(body.as_ref(), initial);
-
-    c.put_object()
-        .bucket(bucket)
-        .key(key)
-        .body(ByteStream::from_static(updated))
-        .if_match(&multipart_etag)
-        .send()
-        .await?;
-
-    let result = c.get_object().bucket(bucket).key(key).send().await?;
-    let body = result.body.collect().await?.into_bytes();
-    assert_eq!(body.as_ref(), updated);
-
-    delete_object(&c, bucket, key).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_put_object_if_match_legacy_md5_fallback() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("if-match-legacy-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    let key = "legacy-object.txt";
-    let initial = b"legacy initial content";
-    let updated = b"legacy updated content";
-
-    create_bucket(&c, bucket).await?;
-
-    let initial_etag = {
-        let result = c
-            .put_object()
-            .bucket(bucket)
-            .key(key)
-            .body(ByteStream::from_static(initial))
-            .send()
-            .await?;
-        result.e_tag().expect("put_object should return e_tag").to_owned()
-    };
-
-    let encode = |s: &str| base64_simd::URL_SAFE_NO_PAD.encode_to_string(s);
-    let internal_info_path =
-        std::path::Path::new(FS_ROOT).join(format!(".bucket-{}.object-{}.internal.json", encode(bucket), encode(key)));
-    fs::remove_file(internal_info_path)?;
-
-    let err = c
-        .put_object()
-        .bucket(bucket)
-        .key(key)
-        .body(ByteStream::from_static(b"wrong overwrite"))
-        .if_match("\"wrong-etag-value\"")
-        .send()
-        .await
-        .expect_err("wrong ETag should fail through MD5 fallback");
-    assert_eq!(err.into_service_error().code(), Some("PreconditionFailed"));
-
-    let result = c.get_object().bucket(bucket).key(key).send().await?;
-    let body = result.body.collect().await?.into_bytes();
-    assert_eq!(body.as_ref(), initial);
-
-    c.put_object()
-        .bucket(bucket)
-        .key(key)
-        .body(ByteStream::from_static(updated))
-        .if_match(&initial_etag)
-        .send()
-        .await?;
-
-    let result = c.get_object().bucket(bucket).key(key).send().await?;
-    let body = result.body.collect().await?.into_bytes();
-    assert_eq!(body.as_ref(), updated);
-
-    delete_object(&c, bucket, key).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_put_object_if_match_rejects_weak_etag() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("if-match-weak-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    let key = "weak-etag.txt";
-    let initial = b"weak etag initial content";
-
-    create_bucket(&c, bucket).await?;
-
-    let initial_etag = {
-        let result = c
-            .put_object()
-            .bucket(bucket)
-            .key(key)
-            .body(ByteStream::from_static(initial))
-            .send()
-            .await?;
-        result.e_tag().expect("put_object should return e_tag").to_owned()
-    };
-    let weak_etag = format!("W/{initial_etag}");
-
-    let err = c
-        .put_object()
-        .bucket(bucket)
-        .key(key)
-        .body(ByteStream::from_static(b"weak overwrite"))
-        .if_match(weak_etag)
-        .send()
-        .await
-        .expect_err("weak ETag must not satisfy If-Match");
-    assert_eq!(err.into_service_error().code(), Some("PreconditionFailed"));
-
-    let result = c.get_object().bucket(bucket).key(key).send().await?;
-    let body = result.body.collect().await?.into_bytes();
-    assert_eq!(body.as_ref(), initial);
-
-    delete_object(&c, bucket, key).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// Regression test for <https://github.com/s3s-project/s3s/issues/67>
-///
-/// `copy_object` should create parent directories when the destination key contains "/"
-#[tokio::test]
-#[tracing::instrument]
-async fn test_copy_object_nested_dst() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-copy-nested-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-
-    create_bucket(&c, bucket).await?;
-
-    // Put a file at the root level
-    let src_key = "source.txt";
-    let content = "copy me into a nested directory";
-    c.put_object()
-        .bucket(bucket)
-        .key(src_key)
-        .body(ByteStream::from_static(content.as_bytes()))
-        .send()
-        .await?;
-
-    // Copy to a nested destination with multiple levels of "/"
-    let dst_key = "deep/nested/path/destination.txt";
-    let copy_source = format!("{bucket}/{src_key}");
-    c.copy_object()
-        .bucket(bucket)
-        .key(dst_key)
-        .copy_source(copy_source)
-        .send()
-        .await?;
-
-    // Verify the copied file exists and has the correct content
-    let ans = c.get_object().bucket(bucket).key(dst_key).send().await?;
-    let body = ans.body.collect().await?.into_bytes();
-    assert_eq!(body.as_ref(), content.as_bytes());
-
-    // Cleanup
-    delete_object(&c, bucket, src_key).await?;
-    delete_object(&c, bucket, dst_key).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// Regression test: `CopyObject` with `src == dst` (self-replace) must
-/// preserve the on-disk content. AWS S3 supports this shape as the
-/// canonical way to update an object's metadata in place. Before the
-/// fix, `tokio::fs::copy(src, dst)` opened the destination with
-/// `O_TRUNC` before reading the source, zeroing the file.
-#[tokio::test]
-#[tracing::instrument]
-async fn test_copy_object_self_replace_preserves_content() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-self-replace-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-
-    create_bucket(&c, bucket).await?;
-
-    let key = "obj.bin";
-    let content = "original content that must survive a self-replace";
-    c.put_object()
-        .bucket(bucket)
-        .key(key)
-        .body(ByteStream::from_static(content.as_bytes()))
-        .send()
-        .await?;
-
-    let before_head = c.head_object().bucket(bucket).key(key).send().await?;
-    let before_last_modified = before_head
-        .last_modified()
-        .expect("head_object should return last_modified")
-        .to_owned();
-
-    tokio::time::sleep(std::time::Duration::from_secs(1)).await;
-
-    let copy_source = format!("{bucket}/{key}");
-    c.copy_object()
-        .bucket(bucket)
-        .key(key)
-        .copy_source(&copy_source)
-        .send()
-        .await?;
-
-    let after_head = c.head_object().bucket(bucket).key(key).send().await?;
-    let after_last_modified = after_head
-        .last_modified()
-        .expect("head_object should return last_modified")
-        .to_owned();
-    assert!(
-        after_last_modified > before_last_modified,
-        "CopyObject self-replace must update LastModified"
-    );
-
-    let got = c.get_object().bucket(bucket).key(key).send().await?;
-    let body = got.body.collect().await?.into_bytes();
-    assert_eq!(body.as_ref(), content.as_bytes(), "CopyObject self-replace must not zero the file");
-
-    delete_object(&c, bucket, key).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// `MetadataDirective: REPLACE` must drop the source metadata and
-/// install the request's metadata + `content_type` on the destination.
-/// Before the fix, `copy_object` ignored both `metadata_directive`
-/// and `metadata` and unconditionally copied the source sidecar
-/// verbatim.
-#[tokio::test]
-#[tracing::instrument]
-async fn test_copy_object_metadata_directive_replace() -> Result<()> {
-    use aws_sdk_s3::types::MetadataDirective;
-
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-meta-replace-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-
-    create_bucket(&c, bucket).await?;
-
-    let src_key = "src.bin";
-    c.put_object()
-        .bucket(bucket)
-        .key(src_key)
-        .body(ByteStream::from_static(b"x"))
-        .content_type("application/octet-stream")
-        .metadata("origin", "v1")
-        .metadata("rev", "1")
-        .metadata("source-only", "keep-out")
-        .send()
-        .await?;
-
-    let dst_key = "dst.bin";
-    let copy_source = format!("{bucket}/{src_key}");
-    c.copy_object()
-        .bucket(bucket)
-        .key(dst_key)
-        .copy_source(&copy_source)
-        .metadata_directive(MetadataDirective::Replace)
-        .content_type("application/pdf")
-        .metadata("origin", "v2")
-        .metadata("rev", "2")
-        .send()
-        .await?;
-
-    let head = c.head_object().bucket(bucket).key(dst_key).send().await?;
-    assert_eq!(
-        head.content_type().unwrap_or(""),
-        "application/pdf",
-        "REPLACE must install the request's content_type on the destination"
-    );
-    let dst_meta = head.metadata().cloned().unwrap_or_default();
-    assert_eq!(dst_meta.get("origin").map(String::as_str), Some("v2"));
-    assert_eq!(dst_meta.get("rev").map(String::as_str), Some("2"));
-    assert_eq!(
-        dst_meta.get("source-only"),
-        None,
-        "REPLACE must drop metadata that only exists on the source object"
-    );
-
-    delete_object(&c, bucket, src_key).await?;
-    delete_object(&c, bucket, dst_key).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// Omitting `MetadataDirective` must use S3's default `COPY` behavior:
-/// propagate source metadata and ignore replacement fields from the request.
-#[tokio::test]
-#[tracing::instrument]
-async fn test_copy_object_metadata_directive_default_copies_source() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-meta-default-copy-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-
-    create_bucket(&c, bucket).await?;
-
-    let src_key = "src.bin";
-    c.put_object()
-        .bucket(bucket)
-        .key(src_key)
-        .body(ByteStream::from_static(b"x"))
-        .content_type("application/octet-stream")
-        .metadata("origin", "v1")
-        .send()
-        .await?;
-
-    let dst_key = "dst.bin";
-    let copy_source = format!("{bucket}/{src_key}");
-    c.copy_object()
-        .bucket(bucket)
-        .key(dst_key)
-        .copy_source(&copy_source)
-        .content_type("application/pdf")
-        .metadata("origin", "v2")
-        .send()
-        .await?;
-
-    let head = c.head_object().bucket(bucket).key(dst_key).send().await?;
-    assert_eq!(
-        head.content_type().unwrap_or(""),
-        "application/octet-stream",
-        "default MetadataDirective must propagate the source content_type"
-    );
-    let dst_meta = head.metadata().cloned().unwrap_or_default();
-    assert_eq!(
-        dst_meta.get("origin").map(String::as_str),
-        Some("v1"),
-        "default MetadataDirective must propagate the source metadata"
-    );
-
-    delete_object(&c, bucket, src_key).await?;
-    delete_object(&c, bucket, dst_key).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// `MetadataDirective: COPY` (the default) must propagate the source
-/// metadata to the destination, ignoring any metadata fields supplied
-/// in the request — exactly as documented at
-/// <https://docs.aws.amazon.com/AmazonS3/latest/API/API_CopyObject.html>.
-#[allow(clippy::doc_markdown)]
-#[tokio::test]
-#[tracing::instrument]
-async fn test_copy_object_metadata_directive_copy_ignores_request_fields() -> Result<()> {
-    use aws_sdk_s3::types::MetadataDirective;
-
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-meta-copy-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-
-    create_bucket(&c, bucket).await?;
-
-    let src_key = "src.bin";
-    c.put_object()
-        .bucket(bucket)
-        .key(src_key)
-        .body(ByteStream::from_static(b"x"))
-        .content_type("application/octet-stream")
-        .metadata("origin", "v1")
-        .send()
-        .await?;
-
-    let dst_key = "dst.bin";
-    let copy_source = format!("{bucket}/{src_key}");
-    c.copy_object()
-        .bucket(bucket)
-        .key(dst_key)
-        .copy_source(&copy_source)
-        .metadata_directive(MetadataDirective::Copy)
-        .content_type("application/pdf") // expected to be ignored under COPY
-        .metadata("origin", "v2")
-        .send()
-        .await?;
-
-    let head = c.head_object().bucket(bucket).key(dst_key).send().await?;
-    assert_eq!(
-        head.content_type().unwrap_or(""),
-        "application/octet-stream",
-        "COPY must propagate the source content_type and ignore the request override"
-    );
-    let dst_meta = head.metadata().cloned().unwrap_or_default();
-    assert_eq!(
-        dst_meta.get("origin").map(String::as_str),
-        Some("v1"),
-        "COPY must propagate the source metadata, not the request fields"
-    );
-
-    delete_object(&c, bucket, src_key).await?;
-    delete_object(&c, bucket, dst_key).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// Test conditional copy with `x-amz-copy-source-if-match`.
-#[tokio::test]
-#[tracing::instrument]
-async fn test_copy_object_if_match() -> Result<()> {
-    use aws_sdk_s3::primitives::DateTime;
-    use aws_sdk_s3::primitives::DateTimeFormat;
-
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-cond-copy-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    let src_key = "source.txt";
-    let content = "conditional copy content";
-    c.put_object()
-        .bucket(bucket)
-        .key(src_key)
-        .body(ByteStream::from_static(content.as_bytes()))
-        .send()
-        .await?;
-
-    let get_result = c.get_object().bucket(bucket).key(src_key).send().await?;
-    let e_tag = get_result.e_tag().expect("get_object should return e_tag").to_owned();
-    let _ = get_result.body.collect().await?;
-
-    let copy_source = format!("{bucket}/{src_key}");
-
-    let dst_key = "dest-match.txt";
-    c.copy_object()
-        .bucket(bucket)
-        .key(dst_key)
-        .copy_source(&copy_source)
-        .copy_source_if_match(&e_tag)
-        .send()
-        .await?;
-
-    let ans = c.get_object().bucket(bucket).key(dst_key).send().await?;
-    let body = ans.body.collect().await?.into_bytes();
-    assert_eq!(body.as_ref(), content.as_bytes());
-
-    let dst_key2 = "dest-match-wildcard.txt";
-    let past = DateTime::from_str("Thu, 01 Jan 2000 00:00:00 GMT", DateTimeFormat::HttpDate)?;
-    c.copy_object()
-        .bucket(bucket)
-        .key(dst_key2)
-        .copy_source(&copy_source)
-        .copy_source_if_match("*")
-        .copy_source_if_unmodified_since(past)
-        .send()
-        .await?;
-
-    let ans = c.get_object().bucket(bucket).key(dst_key2).send().await?;
-    let body = ans.body.collect().await?.into_bytes();
-    assert_eq!(body.as_ref(), content.as_bytes());
-
-    let dst_key3 = "dest-nomatch.txt";
-    let err = c
-        .copy_object()
-        .bucket(bucket)
-        .key(dst_key3)
-        .copy_source(&copy_source)
-        .copy_source_if_match("\"nonexistent-etag\"")
-        .send()
-        .await
-        .expect_err("Expected copy with non-matching If-Match to fail");
-    let service_err = err.into_service_error();
-    assert_eq!(service_err.code(), Some("PreconditionFailed"));
-
-    delete_object(&c, bucket, src_key).await?;
-    delete_object(&c, bucket, dst_key).await?;
-    delete_object(&c, bucket, dst_key2).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// Test conditional copy with `x-amz-copy-source-if-none-match`.
-#[tokio::test]
-#[tracing::instrument]
-async fn test_copy_object_if_none_match() -> Result<()> {
-    use aws_sdk_s3::primitives::DateTime;
-    use aws_sdk_s3::primitives::DateTimeFormat;
-
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-cond-copy-nm-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    let src_key = "source.txt";
-    let content = "conditional copy none match";
-    c.put_object()
-        .bucket(bucket)
-        .key(src_key)
-        .body(ByteStream::from_static(content.as_bytes()))
-        .send()
-        .await?;
-
-    let get_result = c.get_object().bucket(bucket).key(src_key).send().await?;
-    let e_tag = get_result.e_tag().expect("get_object should return e_tag").to_owned();
-    let _ = get_result.body.collect().await?;
-
-    let copy_source = format!("{bucket}/{src_key}");
-
-    let dst_key = "dest-none-match-ok.txt";
-    c.copy_object()
-        .bucket(bucket)
-        .key(dst_key)
-        .copy_source(&copy_source)
-        .copy_source_if_none_match("\"different-etag\"")
-        .send()
-        .await?;
-
-    let ans = c.get_object().bucket(bucket).key(dst_key).send().await?;
-    let body = ans.body.collect().await?.into_bytes();
-    assert_eq!(body.as_ref(), content.as_bytes());
-
-    let dst_key2 = "dest-none-match-fail.txt";
-    let err = c
-        .copy_object()
-        .bucket(bucket)
-        .key(dst_key2)
-        .copy_source(&copy_source)
-        .copy_source_if_none_match(&e_tag)
-        .send()
-        .await
-        .expect_err("Expected copy with matching If-None-Match to fail");
-    let service_err = err.into_service_error();
-    assert_eq!(service_err.code(), Some("PreconditionFailed"));
-
-    let dst_key3 = "dest-none-match-wildcard.txt";
-    let err = c
-        .copy_object()
-        .bucket(bucket)
-        .key(dst_key3)
-        .copy_source(&copy_source)
-        .copy_source_if_none_match("*")
-        .send()
-        .await
-        .expect_err("Expected copy with wildcard If-None-Match to fail for existing source");
-    let service_err = err.into_service_error();
-    assert_eq!(service_err.code(), Some("PreconditionFailed"));
-
-    let dst_key4 = "dest-none-match-precedence.txt";
-    let future = DateTime::from_str("Thu, 01 Jan 2099 00:00:00 GMT", DateTimeFormat::HttpDate)?;
-    c.copy_object()
-        .bucket(bucket)
-        .key(dst_key4)
-        .copy_source(&copy_source)
-        .copy_source_if_none_match("\"different-etag\"")
-        .copy_source_if_modified_since(future)
-        .send()
-        .await?;
-
-    let ans = c.get_object().bucket(bucket).key(dst_key4).send().await?;
-    let body = ans.body.collect().await?.into_bytes();
-    assert_eq!(body.as_ref(), content.as_bytes());
-
-    delete_object(&c, bucket, src_key).await?;
-    delete_object(&c, bucket, dst_key).await?;
-    delete_object(&c, bucket, dst_key4).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// Test conditional copy with `x-amz-copy-source-if-modified-since` and
-/// `x-amz-copy-source-if-unmodified-since`.
-#[tokio::test]
-#[tracing::instrument]
-async fn test_copy_object_if_modified_since() -> Result<()> {
-    use aws_sdk_s3::primitives::DateTime;
-    use aws_sdk_s3::primitives::DateTimeFormat;
-
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-cond-copy-ts-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    let src_key = "source.txt";
-    let content = "conditional copy timestamp";
-    c.put_object()
-        .bucket(bucket)
-        .key(src_key)
-        .body(ByteStream::from_static(content.as_bytes()))
-        .send()
-        .await?;
-
-    let copy_source = format!("{bucket}/{src_key}");
-
-    let dst_key = "dest-modified-ok.txt";
-    let past = DateTime::from_str("Thu, 01 Jan 2000 00:00:00 GMT", DateTimeFormat::HttpDate)?;
-    c.copy_object()
-        .bucket(bucket)
-        .key(dst_key)
-        .copy_source(&copy_source)
-        .copy_source_if_modified_since(past)
-        .send()
-        .await?;
-
-    let ans = c.get_object().bucket(bucket).key(dst_key).send().await?;
-    let body = ans.body.collect().await?.into_bytes();
-    assert_eq!(body.as_ref(), content.as_bytes());
-
-    let dst_key2 = "dest-modified-fail.txt";
-    let future = DateTime::from_str("Thu, 01 Jan 2099 00:00:00 GMT", DateTimeFormat::HttpDate)?;
-    let err = c
-        .copy_object()
-        .bucket(bucket)
-        .key(dst_key2)
-        .copy_source(&copy_source)
-        .copy_source_if_modified_since(future)
-        .send()
-        .await
-        .expect_err("Expected copy with future if-modified-since to fail");
-    let service_err = err.into_service_error();
-    assert_eq!(service_err.code(), Some("PreconditionFailed"));
-
-    let dst_key3 = "dest-unmodified-ok.txt";
-    let future = DateTime::from_str("Thu, 01 Jan 2099 00:00:00 GMT", DateTimeFormat::HttpDate)?;
-    c.copy_object()
-        .bucket(bucket)
-        .key(dst_key3)
-        .copy_source(&copy_source)
-        .copy_source_if_unmodified_since(future)
-        .send()
-        .await?;
-
-    let ans = c.get_object().bucket(bucket).key(dst_key3).send().await?;
-    let body = ans.body.collect().await?.into_bytes();
-    assert_eq!(body.as_ref(), content.as_bytes());
-
-    let dst_key4 = "dest-unmodified-fail.txt";
-    let past = DateTime::from_str("Thu, 01 Jan 2000 00:00:00 GMT", DateTimeFormat::HttpDate)?;
-    let err = c
-        .copy_object()
-        .bucket(bucket)
-        .key(dst_key4)
-        .copy_source(&copy_source)
-        .copy_source_if_unmodified_since(past)
-        .send()
-        .await
-        .expect_err("Expected copy with past if-unmodified-since to fail");
-    let service_err = err.into_service_error();
-    assert_eq!(service_err.code(), Some("PreconditionFailed"));
-
-    delete_object(&c, bucket, src_key).await?;
-    delete_object(&c, bucket, dst_key).await?;
-    delete_object(&c, bucket, dst_key3).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// Test conditional copy against a multipart source object's persisted `ETag`.
-#[tokio::test]
-#[tracing::instrument]
-async fn test_copy_object_conditional_with_multipart_source_etag() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-cond-copy-multipart-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    let src_key = "source-multipart.txt";
-    let content = "multipart conditional copy content";
-
-    let upload_id = c
-        .create_multipart_upload()
-        .bucket(bucket)
-        .key(src_key)
-        .send()
-        .await?
-        .upload_id
-        .expect("create_multipart_upload should return upload_id");
-
-    let upload_result = c
-        .upload_part()
-        .bucket(bucket)
-        .key(src_key)
-        .upload_id(&upload_id)
-        .body(ByteStream::from_static(content.as_bytes()))
-        .part_number(1)
-        .send()
-        .await?;
-
-    let upload = CompletedMultipartUpload::builder()
-        .set_parts(Some(vec![
-            CompletedPart::builder()
-                .e_tag(upload_result.e_tag.expect("upload_part should return e_tag"))
-                .part_number(1)
-                .build(),
-        ]))
-        .build();
-
-    let complete_result = c
-        .complete_multipart_upload()
-        .bucket(bucket)
-        .key(src_key)
-        .multipart_upload(upload)
-        .upload_id(&upload_id)
-        .send()
-        .await?;
-    let multipart_e_tag = complete_result
-        .e_tag()
-        .expect("complete_multipart_upload should return e_tag")
-        .to_owned();
-
-    let copy_source = format!("{bucket}/{src_key}");
-
-    let dst_key = "dest-multipart-match.txt";
-    c.copy_object()
-        .bucket(bucket)
-        .key(dst_key)
-        .copy_source(&copy_source)
-        .copy_source_if_match(&multipart_e_tag)
-        .send()
-        .await?;
-
-    let ans = c.get_object().bucket(bucket).key(dst_key).send().await?;
-    let body = ans.body.collect().await?.into_bytes();
-    assert_eq!(body.as_ref(), content.as_bytes());
-
-    // The destination ETag should match the source multipart ETag (format preserved during copy).
-    let head = c.head_object().bucket(bucket).key(dst_key).send().await?;
-    let dst_etag = head.e_tag().expect("head_object should return e_tag");
-    assert_eq!(
-        dst_etag, multipart_e_tag,
-        "destination ETag should match source multipart ETag after copy"
-    );
-
-    let dst_key2 = "dest-multipart-none-match.txt";
-    let err = c
-        .copy_object()
-        .bucket(bucket)
-        .key(dst_key2)
-        .copy_source(&copy_source)
-        .copy_source_if_none_match(&multipart_e_tag)
-        .send()
-        .await
-        .expect_err("Expected matching multipart ETag to fail If-None-Match");
-    let service_err = err.into_service_error();
-    assert_eq!(service_err.code(), Some("PreconditionFailed"));
-
-    delete_object(&c, bucket, src_key).await?;
-    delete_object(&c, bucket, dst_key).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// Regression test for <https://github.com/s3s-project/s3s/issues/112>
-///
-/// `list_objects_v2` prefix matching should use string-based matching (not `Path::starts_with`)
-/// and `start_after` should work correctly
-#[tokio::test]
-#[tracing::instrument]
-async fn test_list_objects_v2_start_after() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-start-after-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    let content = "test content";
-    let keys = ["aaa.txt", "bbb.txt", "ccc.txt", "ddd.txt"];
-    for key in &keys {
-        c.put_object()
-            .bucket(bucket)
-            .key(*key)
-            .body(ByteStream::from_static(content.as_bytes()))
-            .send()
-            .await?;
-    }
-
-    // start_after="bbb.txt" should return only ccc.txt and ddd.txt
-    let result = c.list_objects_v2().bucket(bucket).start_after("bbb.txt").send().await?;
-
-    let contents: Vec<_> = result.contents().iter().filter_map(|obj| obj.key()).collect();
-    assert_eq!(contents, vec!["ccc.txt", "ddd.txt"]);
-
-    // Cleanup
-    for key in &keys {
-        delete_object(&c, bucket, key).await?;
-    }
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_list_objects_v2_continuation_token_with_delimiter() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-continuation-token-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    let content = "test";
-    let keys = ["a.txt", "dir/1.txt", "z.txt"];
-    for key in &keys {
-        c.put_object()
-            .bucket(bucket)
-            .key(*key)
-            .body(ByteStream::from_static(content.as_bytes()))
-            .send()
-            .await?;
-    }
-
-    let page1 = c.list_objects_v2().bucket(bucket).delimiter("/").max_keys(2).send().await?;
-    assert_eq!(page1.is_truncated(), Some(true));
-    assert_eq!(page1.contents().iter().filter_map(|obj| obj.key()).collect::<Vec<_>>(), vec!["a.txt"]);
-    assert_eq!(
-        page1
-            .common_prefixes()
-            .iter()
-            .filter_map(|prefix| prefix.prefix())
-            .collect::<Vec<_>>(),
-        vec!["dir/"]
-    );
-    let next_token = page1
-        .next_continuation_token()
-        .expect("truncated v2 listing should include next_continuation_token");
-
-    let page2 = c
-        .list_objects_v2()
-        .bucket(bucket)
-        .delimiter("/")
-        .max_keys(2)
-        .continuation_token(next_token)
-        .send()
-        .await?;
-
-    assert_eq!(page2.is_truncated(), Some(false));
-    assert_eq!(page2.contents().iter().filter_map(|obj| obj.key()).collect::<Vec<_>>(), vec!["z.txt"]);
-    assert!(
-        page2.common_prefixes().is_empty(),
-        "continuation token should advance past the previous common prefix"
-    );
-
-    for key in &keys {
-        delete_object(&c, bucket, key).await?;
-    }
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// Regression test for <https://github.com/s3s-project/s3s/issues/112>
-///
-/// Prefix matching must use string comparison, not `Path::starts_with` which is stricter.
-/// For example, prefix "dir/sub" should match key "dir/subdir/file.txt".
-#[tokio::test]
-#[tracing::instrument]
-async fn test_list_objects_v2_prefix_string_matching() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-prefix-match-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    let content = "test";
-    let keys = ["dir/subdir/file1.txt", "dir/subother/file2.txt", "dir/other/file3.txt"];
-    for key in &keys {
-        c.put_object()
-            .bucket(bucket)
-            .key(*key)
-            .body(ByteStream::from_static(content.as_bytes()))
-            .send()
-            .await?;
-    }
-
-    // Prefix "dir/sub" should match "dir/subdir/..." and "dir/subother/..."
-    // but NOT "dir/other/..."
-    // Path::starts_with would fail here because it requires component boundaries
-    let result = c.list_objects_v2().bucket(bucket).prefix("dir/sub").send().await?;
-
-    let contents: Vec<_> = result.contents().iter().filter_map(|obj| obj.key()).collect();
-    assert_eq!(contents.len(), 2, "Expected 2 objects matching prefix 'dir/sub', got {contents:?}");
-    assert!(contents.contains(&"dir/subdir/file1.txt"));
-    assert!(contents.contains(&"dir/subother/file2.txt"));
-
-    // Cleanup
-    for key in &keys {
-        delete_object(&c, bucket, key).await?;
-    }
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// Regression test for <https://github.com/s3s-project/s3s/issues/116>
-///
-/// `put_object` should write atomically via a temp file to prevent incomplete writes.
-/// Verify that the file is fully written and readable after `put_object` completes.
-#[tokio::test]
-#[tracing::instrument]
-async fn test_put_object_atomic_write() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-atomic-write-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    // Write a reasonably sized object
-    let content = "x".repeat(1024 * 64); // 64 KB
-    let key = "atomic-test.bin";
-
-    c.put_object()
-        .bucket(bucket)
-        .key(key)
-        .body(ByteStream::from(content.clone().into_bytes()))
-        .send()
-        .await?;
-
-    // Read it back immediately and verify full content
-    let ans = c.get_object().bucket(bucket).key(key).send().await?;
-    let body = ans.body.collect().await?.into_bytes();
-    assert_eq!(body.len(), content.len(), "Content length mismatch");
-    assert_eq!(body.as_ref(), content.as_bytes(), "Content mismatch");
-
-    // Verify no temp files remain in the FS root
-    let entries: Vec<_> = fs::read_dir(FS_ROOT)?
-        .filter_map(Result::ok)
-        .filter(|e| {
-            let name = e.file_name();
-            let name = name.to_str().unwrap_or("");
-            name.starts_with(".tmp.") && name.ends_with(".internal.part")
-        })
-        .collect();
-    assert!(entries.is_empty(), "Leftover temp files found: {entries:?}");
-
-    // Cleanup
-    delete_object(&c, bucket, key).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// Regression test for <https://github.com/s3s-project/s3s/issues/51>
-///
-/// Multipart `upload_id` should be bound to the credentials that created it.
-/// A different user should not be able to upload parts or complete the upload.
-#[tokio::test]
-#[tracing::instrument]
-#[allow(clippy::too_many_lines)]
-async fn test_multipart_upload_id_auth() -> Result<()> {
-    let _guard = serial().await;
-
-    // Create a service with two sets of credentials
-    let cred_user1 = Credentials::new("AKUSER1EXAMPLE", "secretkey1example", None, None, "user1");
-    let cred_user2 = Credentials::new("AKUSER2EXAMPLE", "secretkey2example", None, None, "user2");
-
-    let mut auth = SimpleAuth::new();
-    auth.register(cred_user1.access_key_id().to_string(), cred_user1.secret_access_key().into());
-    auth.register(cred_user2.access_key_id().to_string(), cred_user2.secret_access_key().into());
-
-    fs::create_dir_all(FS_ROOT).unwrap();
-    let fs = FileSystem::new(FS_ROOT).unwrap();
-    let service = {
-        let mut b = S3ServiceBuilder::new(fs);
-        b.set_auth(auth);
-        b.set_host(SingleDomain::new(DOMAIN_NAME).unwrap());
-        b.build()
-    };
-
-    // Create client for user1
-    let config_user1 = SdkConfig::builder()
-        .credentials_provider(SharedCredentialsProvider::new(cred_user1.clone()))
-        .http_client(s3s_aws::Client::from(service.clone()))
-        .region(Region::new(REGION))
-        .endpoint_url(format!("http://{DOMAIN_NAME}"))
-        .build();
-    let c1 = Client::new(&config_user1);
-
-    // Create client for user2
-    let config_user2 = SdkConfig::builder()
-        .credentials_provider(SharedCredentialsProvider::new(cred_user2))
-        .http_client(s3s_aws::Client::from(service))
-        .region(Region::new(REGION))
-        .endpoint_url(format!("http://{DOMAIN_NAME}"))
-        .build();
-    let c2 = Client::new(&config_user2);
-
-    let bucket = format!("test-multipart-auth-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    let key = "auth-test.txt";
-
-    // User1 creates bucket and starts multipart upload
-    create_bucket(&c1, bucket).await?;
-
-    let upload_id = {
-        let ans = c1.create_multipart_upload().bucket(bucket).key(key).send().await?;
-        ans.upload_id.unwrap()
-    };
-    let upload_id = upload_id.as_str();
-
-    // User2 tries to upload a part - should fail with AccessDenied
-    let result = c2
-        .upload_part()
-        .bucket(bucket)
-        .key(key)
-        .upload_id(upload_id)
-        .body(ByteStream::from_static(b"unauthorized part"))
-        .part_number(1)
-        .send()
-        .await;
-
-    let err = result.expect_err("Expected AccessDenied when user2 tries to upload part");
-    let service_err = err.into_service_error();
-    assert_eq!(
-        service_err.code(),
-        Some("AccessDenied"),
-        "Expected AccessDenied error code, got: {:?}",
-        service_err.code()
-    );
-
-    // User1 should be able to upload a part
-    let upload_parts = {
-        let body = ByteStream::from_static(b"authorized part");
-        let ans = c1
-            .upload_part()
-            .bucket(bucket)
-            .key(key)
-            .upload_id(upload_id)
-            .body(body)
-            .part_number(1)
-            .send()
-            .await?;
-
-        vec![
-            CompletedPart::builder()
-                .e_tag(ans.e_tag.unwrap_or_default())
-                .part_number(1)
-                .build(),
-        ]
-    };
-
-    // User2 tries to complete the upload - should fail with AccessDenied
-    let upload = CompletedMultipartUpload::builder()
-        .set_parts(Some(upload_parts.clone()))
-        .build();
-    let result = c2
-        .complete_multipart_upload()
-        .bucket(bucket)
-        .key(key)
-        .multipart_upload(upload)
-        .upload_id(upload_id)
-        .send()
-        .await;
-
-    let err = result.expect_err("Expected AccessDenied when user2 tries to complete upload");
-    let service_err = err.into_service_error();
-    assert_eq!(
-        service_err.code(),
-        Some("AccessDenied"),
-        "Expected AccessDenied error code, got: {:?}",
-        service_err.code()
-    );
-
-    // User1 completes the upload
-    let upload = CompletedMultipartUpload::builder().set_parts(Some(upload_parts)).build();
-    c1.complete_multipart_upload()
-        .bucket(bucket)
-        .key(key)
-        .multipart_upload(upload)
-        .upload_id(upload_id)
-        .send()
-        .await?;
-
-    // Cleanup
-    delete_object(&c1, bucket, key).await?;
-    delete_bucket(&c1, bucket).await?;
-
-    Ok(())
-}
-
-/// Test that `CompleteMultipartUpload` with `If-None-Match: *` succeeds when object doesn't exist
-/// and fails with `PreconditionFailed` (412) when object already exists.
-#[tokio::test]
-#[tracing::instrument]
-async fn test_complete_multipart_if_none_match() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-cmu-ifnm-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    let key = "multipart-conditional.txt";
-    let content = "abcdefghijklmnopqrstuvwxyz/0123456789/!@#$%^&*();\n";
-
-    create_bucket(&c, bucket).await?;
-
-    // Test 1: CompleteMultipartUpload with If-None-Match: * should succeed when object doesn't exist
-    debug!("Test 1: CompleteMultipartUpload with If-None-Match: * on non-existent object");
-    {
-        let (upload_id, upload_parts) = do_multipart_upload(&c, bucket, key, content.as_bytes()).await?;
-        let upload = CompletedMultipartUpload::builder().set_parts(Some(upload_parts)).build();
-
-        let result = c
-            .complete_multipart_upload()
-            .bucket(bucket)
-            .key(key)
-            .upload_id(&upload_id)
-            .multipart_upload(upload)
-            .if_none_match("*")
-            .send()
-            .await;
-
-        match result {
-            Ok(_) => debug!("✓ Successfully completed multipart upload with If-None-Match: *"),
-            Err(e) => panic!(
-                "Expected CompleteMultipartUpload with If-None-Match: * to succeed when object doesn't exist, but got error: {e:?}"
-            ),
-        }
-    }
-
-    // Verify the object was created
-    {
-        let result = c.get_object().bucket(bucket).key(key).send().await?;
-        let body = result.body.collect().await?.into_bytes();
-        assert_eq!(body.as_ref(), content.as_bytes());
-        debug!("✓ Verified object was created via multipart upload");
-    }
-
-    // Test 2: CompleteMultipartUpload with If-None-Match: * should fail when object exists
-    debug!("Test 2: CompleteMultipartUpload with If-None-Match: * on existing object");
-    {
-        let (upload_id, upload_parts) = do_multipart_upload(&c, bucket, key, b"new content").await?;
-        let upload = CompletedMultipartUpload::builder().set_parts(Some(upload_parts)).build();
-
-        let result = c
-            .complete_multipart_upload()
-            .bucket(bucket)
-            .key(key)
-            .upload_id(&upload_id)
-            .multipart_upload(upload)
-            .if_none_match("*")
-            .send()
-            .await;
-
-        match result {
-            Ok(_) => {
-                panic!("Expected CompleteMultipartUpload with If-None-Match: * to fail when object exists, but it succeeded")
-            }
-            Err(e) => {
-                let service_err = e.into_service_error();
-                debug!("✓ Expected error when object exists: {service_err:?}");
-                assert_eq!(
-                    service_err.code(),
-                    Some("PreconditionFailed"),
-                    "Expected PreconditionFailed, got: {:?}",
-                    service_err.code()
-                );
-            }
-        }
-    }
-
-    // Verify the object wasn't overwritten
-    {
-        let result = c.get_object().bucket(bucket).key(key).send().await?;
-        let body = result.body.collect().await?.into_bytes();
-        assert_eq!(body.as_ref(), content.as_bytes());
-        debug!("✓ Verified object was not overwritten");
-    }
-
-    // Cleanup
-    delete_object(&c, bucket, key).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// Test that `CompleteMultipartUpload` with `If-Match` succeeds when `ETag` matches
-/// and fails with `PreconditionFailed` (412) when `ETag` doesn't match or object is absent.
-#[tokio::test]
-#[tracing::instrument]
-#[allow(clippy::too_many_lines)]
-async fn test_complete_multipart_if_match() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-cmu-ifm-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    let key = "multipart-conditional-match.txt";
-    let content = "abcdefghijklmnopqrstuvwxyz/0123456789/!@#$%^&*();\n";
-
-    create_bucket(&c, bucket).await?;
-
-    // Test 1: CompleteMultipartUpload with If-Match on absent object should fail
-    debug!("Test 1: CompleteMultipartUpload with If-Match on absent object");
-    {
-        let (upload_id, upload_parts) = do_multipart_upload(&c, bucket, key, b"some content").await?;
-        let upload = CompletedMultipartUpload::builder()
-            .set_parts(Some(upload_parts.clone()))
-            .build();
-
-        let result = c
-            .complete_multipart_upload()
-            .bucket(bucket)
-            .key(key)
-            .upload_id(&upload_id)
-            .multipart_upload(upload)
-            .if_match("\"some-etag\"")
-            .send()
-            .await;
-
-        match result {
-            Ok(_) => panic!("Expected CompleteMultipartUpload with If-Match on absent object to fail, but it succeeded"),
-            Err(e) => {
-                let service_err = e.into_service_error();
-                debug!("✓ Expected error on absent object: {service_err:?}");
-                assert_eq!(
-                    service_err.code(),
-                    Some("PreconditionFailed"),
-                    "Expected PreconditionFailed, got: {:?}",
-                    service_err.code()
-                );
-            }
-        }
-
-        // Verify the upload was not consumed: retry the same upload_id should still fail with precondition
-        let upload2 = CompletedMultipartUpload::builder().set_parts(Some(upload_parts)).build();
-        let result2 = c
-            .complete_multipart_upload()
-            .bucket(bucket)
-            .key(key)
-            .upload_id(&upload_id)
-            .multipart_upload(upload2)
-            .if_match("\"some-etag\"")
-            .send()
-            .await;
-
-        let err2 = result2.expect_err("Expected retry to also fail");
-        let service_err2 = err2.into_service_error();
-        assert_eq!(
-            service_err2.code(),
-            Some("PreconditionFailed"),
-            "Expected retry to fail with PreconditionFailed, got: {:?}",
-            service_err2.code()
-        );
-        debug!("✓ Upload was not consumed by failed precondition check");
-    }
-
-    // Create the object with a known ETag via put_object
-    let initial_etag = {
-        let body = ByteStream::from_static(content.as_bytes());
-        let result = c.put_object().bucket(bucket).key(key).body(body).send().await?;
-        result.e_tag().unwrap().to_owned()
-    };
-    debug!("Initial ETag: {initial_etag}");
-
-    // Test 2: CompleteMultipartUpload with If-Match and wrong ETag should fail
-    debug!("Test 2: CompleteMultipartUpload with If-Match and wrong ETag");
-    {
-        let (upload_id, upload_parts) = do_multipart_upload(&c, bucket, key, b"new content").await?;
-        let upload = CompletedMultipartUpload::builder().set_parts(Some(upload_parts)).build();
-
-        let result = c
-            .complete_multipart_upload()
-            .bucket(bucket)
-            .key(key)
-            .upload_id(&upload_id)
-            .multipart_upload(upload)
-            .if_match("\"wrong-etag-value\"")
-            .send()
-            .await;
-
-        match result {
-            Ok(_) => panic!("Expected CompleteMultipartUpload with wrong If-Match to fail, but it succeeded"),
-            Err(e) => {
-                let service_err = e.into_service_error();
-                debug!("✓ Expected error with wrong ETag: {service_err:?}");
-                assert_eq!(
-                    service_err.code(),
-                    Some("PreconditionFailed"),
-                    "Expected PreconditionFailed, got: {:?}",
-                    service_err.code()
-                );
-            }
-        }
-    }
-
-    // Verify the object wasn't overwritten
-    {
-        let result = c.get_object().bucket(bucket).key(key).send().await?;
-        let body = result.body.collect().await?.into_bytes();
-        assert_eq!(body.as_ref(), content.as_bytes());
-        debug!("✓ Verified object was not overwritten");
-    }
-
-    // Test 3: CompleteMultipartUpload with If-Match and correct ETag should succeed
-    debug!("Test 3: CompleteMultipartUpload with If-Match and correct ETag");
-    {
-        let new_content = b"updated via conditional multipart";
-        let (upload_id, upload_parts) = do_multipart_upload(&c, bucket, key, new_content).await?;
-        let upload = CompletedMultipartUpload::builder().set_parts(Some(upload_parts)).build();
-
-        let result = c
-            .complete_multipart_upload()
-            .bucket(bucket)
-            .key(key)
-            .upload_id(&upload_id)
-            .multipart_upload(upload)
-            .if_match(&initial_etag)
-            .send()
-            .await;
-
-        match result {
-            Ok(_) => debug!("✓ Successfully completed multipart upload with matching If-Match"),
-            Err(e) => panic!("Expected CompleteMultipartUpload with matching If-Match to succeed, but got error: {e:?}"),
-        }
-    }
-
-    // Verify the object was updated (use head_object to avoid body checksum issues
-    // since complete_multipart_upload doesn't update internal checksum info)
-    {
-        let result = c.head_object().bucket(bucket).key(key).send().await?;
-        assert!(result.content_length().is_some());
-        debug!("✓ Verified object exists after conditional multipart upload");
-    }
-
-    // Cleanup
-    delete_object(&c, bucket, key).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// Test that `CompleteMultipartUpload` with `If-Match: *` succeeds when object exists
-/// and fails with `PreconditionFailed` (412) when object is absent.
-#[tokio::test]
-#[tracing::instrument]
-async fn test_complete_multipart_if_match_wildcard() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-cmu-ifm-wc-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    let key = "multipart-conditional-match-wildcard.txt";
-
-    create_bucket(&c, bucket).await?;
-
-    // Test 1: If-Match: * on absent object should fail
-    debug!("Test 1: CompleteMultipartUpload with If-Match: * on absent object");
-    {
-        let (upload_id, upload_parts) = do_multipart_upload(&c, bucket, key, b"some content").await?;
-        let upload = CompletedMultipartUpload::builder().set_parts(Some(upload_parts)).build();
-
-        let err = c
-            .complete_multipart_upload()
-            .bucket(bucket)
-            .key(key)
-            .upload_id(&upload_id)
-            .multipart_upload(upload)
-            .if_match("*")
-            .send()
-            .await
-            .expect_err("Expected If-Match: * on absent object to fail");
-
-        let service_err = err.into_service_error();
-        assert_eq!(
-            service_err.code(),
-            Some("PreconditionFailed"),
-            "Expected PreconditionFailed, got: {:?}",
-            service_err.code()
-        );
-        debug!("✓ If-Match: * correctly rejected absent object");
-    }
-
-    // Create the object
-    c.put_object()
-        .bucket(bucket)
-        .key(key)
-        .body(ByteStream::from_static(b"existing object"))
-        .send()
-        .await?;
-
-    // Test 2: If-Match: * on existing object should succeed
-    debug!("Test 2: CompleteMultipartUpload with If-Match: * on existing object");
-    {
-        let (upload_id, upload_parts) = do_multipart_upload(&c, bucket, key, b"new content").await?;
-        let upload = CompletedMultipartUpload::builder().set_parts(Some(upload_parts)).build();
-
-        c.complete_multipart_upload()
-            .bucket(bucket)
-            .key(key)
-            .upload_id(&upload_id)
-            .multipart_upload(upload)
-            .if_match("*")
-            .send()
-            .await
-            .expect("Expected If-Match: * on existing object to succeed");
-
-        debug!("✓ If-Match: * correctly accepted existing object");
-    }
-
-    // Cleanup
-    delete_object(&c, bucket, key).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_list_objects_v2_continuation_token_pagination() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-continuation-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    let keys = ["a.txt", "b.txt", "c.txt", "d.txt", "e.txt"];
-    for key in &keys {
-        c.put_object()
-            .bucket(bucket)
-            .key(*key)
-            .body(ByteStream::from_static(b"x"))
-            .send()
-            .await?;
-    }
-
-    // Walk all pages with max_keys=2
-    let mut all_keys: Vec<String> = Vec::new();
-    let mut token: Option<String> = None;
-    loop {
-        let mut req = c.list_objects_v2().bucket(bucket).max_keys(2);
-        if let Some(t) = &token {
-            req = req.continuation_token(t.clone());
-        }
-        let page = req.send().await?;
-
-        all_keys.extend(page.contents().iter().filter_map(|o| o.key().map(String::from)));
-
-        if page.is_truncated() != Some(true) {
-            break;
-        }
-        token = page.next_continuation_token().map(String::from);
-        assert!(token.is_some(), "is_truncated is true but next_continuation_token is missing");
-    }
-
-    assert_eq!(all_keys, vec!["a.txt", "b.txt", "c.txt", "d.txt", "e.txt"]);
-
-    // Cleanup
-    for key in &keys {
-        delete_object(&c, bucket, key).await?;
-    }
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// When both `continuation_token` and `start_after` are present, the stricter
-/// (larger) bound wins so we never re-list keys the caller already skipped.
-#[tokio::test]
-#[tracing::instrument]
-async fn test_list_objects_v2_continuation_token_and_start_after_uses_max() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-ct-max-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    let keys = ["a.txt", "b.txt", "c.txt", "d.txt", "e.txt"];
-    for key in &keys {
-        c.put_object()
-            .bucket(bucket)
-            .key(*key)
-            .body(ByteStream::from_static(b"x"))
-            .send()
-            .await?;
-    }
-
-    // start_after="d.txt" is larger than continuation_token="b.txt", so we resume after d.txt
-    let result = c
-        .list_objects_v2()
-        .bucket(bucket)
-        .continuation_token("b.txt")
-        .start_after("d.txt")
-        .send()
-        .await?;
-    let result_keys: Vec<_> = result.contents().iter().filter_map(|o| o.key()).collect();
-    assert_eq!(result_keys, vec!["e.txt"], "should resume after the larger value (d.txt)");
-
-    // continuation_token="c.txt" is larger than start_after="a.txt", so we resume after c.txt
-    let result = c
-        .list_objects_v2()
-        .bucket(bucket)
-        .continuation_token("c.txt")
-        .start_after("a.txt")
-        .send()
-        .await?;
-    let result_keys: Vec<_> = result.contents().iter().filter_map(|o| o.key()).collect();
-    assert_eq!(result_keys, vec!["d.txt", "e.txt"], "should resume after the larger value (c.txt)");
-
-    // Cleanup
-    for key in &keys {
-        delete_object(&c, bucket, key).await?;
-    }
-
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-/// `max_keys=0` return `is_truncated=false` with no continuation token
-#[tokio::test]
-#[tracing::instrument]
-async fn test_list_objects_v2_max_keys_zero() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-max-keys-zero-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    let keys = ["a.txt", "b.txt", "c.txt"];
-    for key in &keys {
-        c.put_object()
-            .bucket(bucket)
-            .key(*key)
-            .body(ByteStream::from_static(b"x"))
-            .send()
-            .await?;
-    }
-
-    let result = c.list_objects_v2().bucket(bucket).max_keys(0).send().await?;
-
-    assert_eq!(result.is_truncated(), Some(false), "max_keys=0 should not be truncated");
-    assert!(
-        result.next_continuation_token().is_none(),
-        "max_keys=0 should not return a continuation token"
-    );
-    assert!(result.contents().is_empty(), "max_keys=0 should return no objects");
-
-    // Cleanup
-    for key in &keys {
-        delete_object(&c, bucket, key).await?;
-    }
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_head_object_no_such_key() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-head-no-such-key-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    let result = c.head_object().bucket(bucket).key("nonexistent-object").send().await;
-    let err = result.expect_err("Expected NoSuchKey for missing object");
-    let service_err = err.into_service_error();
-    assert_eq!(service_err.code(), Some("NoSuchKey"), "Expected NoSuchKey, got: {:?}", service_err.code());
-
-    delete_bucket(&c, bucket).await?;
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_head_object_directory_prefix_returns_no_such_key() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-head-directory-prefix-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    let key = "prefix/object.txt";
-    create_bucket(&c, bucket).await?;
-    c.put_object()
-        .bucket(bucket)
-        .key(key)
-        .body(ByteStream::from_static(b"content"))
-        .send()
-        .await?;
-
-    let result = c.head_object().bucket(bucket).key("prefix").send().await;
-    let err = result.expect_err("Expected NoSuchKey for a directory-like prefix");
-    let service_err = err.into_service_error();
-    assert_eq!(service_err.code(), Some("NoSuchKey"), "Expected NoSuchKey, got: {:?}", service_err.code());
-
-    delete_object(&c, bucket, key).await?;
-    delete_bucket(&c, bucket).await?;
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_head_object_no_such_bucket() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-head-no-such-bucket-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-
-    let result = c.head_object().bucket(bucket).key("some-key").send().await;
-    let err = result.expect_err("Expected NoSuchBucket for missing bucket");
-    let service_err = err.into_service_error();
-    assert_eq!(
-        service_err.code(),
-        Some("NoSuchBucket"),
-        "Expected NoSuchBucket, got: {:?}",
-        service_err.code()
-    );
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_upload_part_copy_empty_source() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-upc-empty-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    create_bucket(&c, bucket).await?;
-
-    let src_key = "empty.txt";
-    c.put_object()
-        .bucket(bucket)
-        .key(src_key)
-        .body(ByteStream::from_static(b""))
-        .send()
-        .await?;
-
-    let dst_key = "dst.txt";
-    let upload_id = c
-        .create_multipart_upload()
-        .bucket(bucket)
-        .key(dst_key)
-        .send()
-        .await?
-        .upload_id
-        .unwrap();
-
-    let copy_source = format!("{bucket}/{src_key}");
-    c.upload_part_copy()
-        .bucket(bucket)
-        .key(dst_key)
-        .copy_source(copy_source)
-        .upload_id(&upload_id)
-        .part_number(1)
-        .send()
-        .await?;
-
-    c.abort_multipart_upload()
-        .bucket(bucket)
-        .key(dst_key)
-        .upload_id(&upload_id)
-        .send()
-        .await?;
-
-    delete_object(&c, bucket, src_key).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
-
-#[tokio::test]
-#[tracing::instrument]
-async fn test_head_object_etag_and_checksum() -> Result<()> {
-    let _guard = serial().await;
-
-    let c = Client::new(config());
-    let bucket = format!("test-head-etag-{}", Uuid::new_v4());
-    let bucket = bucket.as_str();
-    let key = "sample.txt";
-    let content = "hello world\n";
-    let crc32c = base64_simd::STANDARD.encode_to_string(crc32c::crc32c(content.as_bytes()).to_be_bytes());
-
-    create_bucket(&c, bucket).await?;
-
-    // Put object with checksum
-    let put_result = c
-        .put_object()
-        .bucket(bucket)
-        .key(key)
-        .body(ByteStream::from_static(content.as_bytes()))
-        .checksum_crc32_c(crc32c.as_str())
-        .send()
-        .await?;
-    let put_e_tag = put_result.e_tag().unwrap().to_owned();
-
-    // Head object and verify e_tag is present and matches put_object
-    let head_result = c
-        .head_object()
-        .bucket(bucket)
-        .key(key)
-        .checksum_mode(ChecksumMode::Enabled)
-        .send()
-        .await?;
-    let head_e_tag = head_result.e_tag().expect("head_object should return e_tag").to_owned();
-    assert_eq!(head_e_tag, put_e_tag, "head_object e_tag should match put_object e_tag");
-
-    // Verify checksum is returned
-    let head_crc32c = head_result
-        .checksum_crc32_c()
-        .expect("head_object should return checksum_crc32c");
-    assert_eq!(head_crc32c, crc32c);
-
-    // Get object and verify e_tag matches
-    let get_result = c
-        .get_object()
-        .bucket(bucket)
-        .key(key)
-        .checksum_mode(ChecksumMode::Enabled)
-        .send()
-        .await?;
-    let get_e_tag = get_result.e_tag().expect("get_object should return e_tag").to_owned();
-    assert_eq!(head_e_tag, get_e_tag, "head_object e_tag should match get_object e_tag");
-
-    // Cleanup
-    delete_object(&c, bucket, key).await?;
-    delete_bucket(&c, bucket).await?;
-
-    Ok(())
-}
